@@ -60,52 +72,61 @@
     }
   },
   "dependencies": {
-    "body-parser": "1.15.2",
-    "check-node-version": "2.0.1",
-    "country-language": "0.1.7",
+    "body-parser": "1.19.0",
+    "check-node-version": "4.0.1",
     "electron-updater": "2.8.9",
-    "express": "4.14.0",
-    "font-awesome": "4.7.0",
-    "flag-icon-css": "2.8.0",
-    "http-shutdown": "1.1.0",
+    "express": "4.17.1",
+    "flag-icon-css": "3.4.5",
+    "http-shutdown": "1.2.1",
     "json-stable-stringify": "1.0.1",
     "json2csv": "3.7.3",
-    "lodash": "4.17.11",
+    "lodash": "4.17.15",
     "node-forge": "0.6.46",
-    "notosans-fontface": "1.0.1",
-    "pm2": "2.10.1",
-    "request": "2.79.0",
-    "request-stats": "2.0.1",
-    "semver": "5.3.0",
-    "@luminati-io/socksv5": "0.0.7",
-    "sqlite3": "4.0.2",
-    "sudo-prompt": "6.2.1",
-    "taskkill": "2.0.0",
-    "tasklist": "3.1.0",
+    "pm2": "4.1.2",
+    "request": "2.88.0",
+    "request-stats": "3.0.0",
+    "semver": "6.3.0",
+    "sudo-prompt": "9.0.0",
+    "taskkill": "3.1.0",
+    "tasklist": "4.0.1",
     "tough-cookie-file-store": "1.2.0",
-    "universal-analytics": "0.4.9",
-    "yargs": "6.5.0",
-    "opn": "5.1.0",
-    "prismjs": "1.8.4",
-    "classnames": "2.2.5",
+    "yargs": "14.2.0",
+    "open": "7.0.0",
+    "prismjs": "1.17.1",
+    "classnames": "2.2.6",
     "netmask": "1.0.6",
-    "rotating-file-stream": "1.3.4",
-    "server-destroy": "1.0.1",
-    "uuid": "3.2.1",
-    "compression": "1.7.2",
-    "json-object-stream": "0.1.1",
-    "ws": "5.1.1",
+    "uuid": "3.3.3",
+    "compression": "1.7.4",
+    "ws": "7.2.0",
     "diacritic": "0.0.2",
-    "underscore": "1.8.3",
-    "node-getopt": "0.2.3",
     "rimraf": "2.5.2",
     "log-buffer": "0.0.3",
-    "cheerio": "1.0.0-rc.2",
-    "puppeteer": "1.8.0",
-    "nightmare": "3.0.1",
-    "prompt-sync": "4.1.6",
-    "lokijs": "1.5.5",
-    "ps-list": "5.0.0"
+    "cheerio": "1.0.0-rc.3",
+    "prompt-sync": "4.1.7",
+    "lum_lokijs": "1.5.11",
+    "lum_socksv5": "0.0.11",
+    "ps-list": "6.3.0",
+    "extract-zip": "1.6.7",
+    "cookie": "0.4.0",
+    "winston": "3.2.1",
+    "winston-daily-rotate-file": "4.2.1",
+    "bootstrap": "3.4.1",
+    "codemirror": "5.49.2",
+    "es6-shim": "0.35.5",
+    "event-target-shim": "5.0.1",
+    "jquery": "3.4.1",
+    "moment": "2.24.0",
+    "react": "16.13.1",
+    "react-dom": "16.13.1",
+    "react-router-dom": "5.1.2",
+    "react-bootstrap": "1.0.0",
+    "react-bootstrap-typeahead": "4.1.0",
+    "react-waypoint": "9.0.2",
+    "react-select": "3.0.8",
+    "react-tooltip": "3.11.1",
+    "react-virtualized": "9.21.2",
+    "file-saver": "2.0.2",
+    "lum_crontab": "1.0.0"
   },
   "description": "A configurable local proxy for luminati.io",
   "devDependencies": {
@@ -118,55 +139,45 @@
     "babel-preset-react": "6.24.1",
     "babel-plugin-transform-object-rest-spread": "6.23.0",
     "babel-plugin-transform-runtime": "6.23.0",
-    "bootstrap": "3.3.7",
-    "codemirror": "5.21.0",
     "css-loader": "0.28.4",
-    "electron": "1.7.1",
+    "electron": "4.0.4",
     "electron-builder": "16.6.1",
-    "electron-rebuild": "1.7.2",
-    "es6-shim": "0.35.3",
-    "file-loader": "0.11.2",
+    "electron-rebuild": "1.8.4",
     "html-webpack-plugin": "2.30.1",
-    "jquery": "3.1.1",
     "less": "2.7.2",
     "less-loader": "4.0.5",
     "lolex": "2.1.1",
-    "mocha": "3.2.0",
-    "moment": "2.17.1",
+    "mocha": "5.2.0",
     "nock": "9.0.11",
-    "react": "16.3.2",
-    "react-bootstrap": "0.31.0",
-    "react-bootstrap-typeahead": "2.0.2",
-    "react-dom": "16.1.1",
-    "react-router-dom": "4.2.2",
     "style-loader": "0.18.2",
-    "swagger-ui": "2.2.8",
-    "url-loader": "0.5.9",
-    "file-saver": "1.3.3",
+    "file-loader": "2.0.0",
+    "url-loader": "1.1.2",
     "webpack": "3.4.0",
-    "recompose": "0.26.0",
-    "prop-types": "15.6.0",
-    "react-waypoint": "8.0.1",
-    "react-select": "2.0.0"
+    "prop-types": "15.6.0"
   },
   "engines": {
-    "node": ">=10.0",
-    "npm": "4.6.1"
+    "node": ">=10.0"
   },
   "homepage": "https://luminati.io/",
+  "api_domain": "lum-lpm.com",
+  "api_domain_fallback": "l-lpm.com",
+  "mail_domain": "luminati-china-mail.co",
   "keywords": [
     "luminati",
     "proxy",
-    "hola"
+    "lpm"
   ],
   "license": "ISC",
   "lpm": {
     "files": [
+      ".npmrc",
       "bin/ca.crt",
       "bin/ca.key",
-      "bin/check_compat.js",
+      "bin/cert_gen.sh",
       "bin/index.js",
-      "bin/lum_common.js",
+      "bin/lpm_downgrade.sh",
+      "bin/lpm_upgrade.sh",
+      "bin/lpm_util.sh",
       "bin/lum_electron.js",
       "bin/lum_electron_index.js",
       "bin/lum_node.js",
