@@ -2,12 +2,14 @@
 'use strict'; /*jslint node:true, esnext:true*/
 const _ = require('lodash');
 const string = require('../util/string.js');
+const lpm_config = require('../util/lpm_config.js');
 const diacritic = require('diacritic');
 const qw = string.qw;
 
 const boolean_part = {
     raw: true,
     direct: true,
+    unblocker: true,
 };
 const abbr = {
     cu: 'customer',
@@ -15,22 +17,18 @@ const abbr = {
     k: 'key',
     d: 'direct',
     s: 'session',
-    dbg: 'debug',
     cy: 'country',
     st: 'state',
     ct: 'city',
+    ub: 'unblocker',
+    unblock: 'unblocker',
 };
 const fields = qw`customer zone country state city session asn dns
-    cid ip raw direct debug mobile vip carrier`;
+    cid ip raw direct mobile vip carrier route_err unblocker debug os`;
 
 // XXX krzysztof: copied from /protocol. try to import and reuse
 const escape_geo = val=>
     diacritic.clean(val).toLowerCase().replace(/[^a-z0-9]/g, '');
-
-// XXX colin/ovidiu: remove this function and us from util if possible
-let short = {};
-for (let key in abbr)
-    short[abbr[key]] = key;
 
 const parse = header=>{
     if (!header)
@@ -38,7 +36,7 @@ const parse = header=>{
     let m = header.match(/^Basic (.*)/);
     if (!m)
         return;
-    header = new Buffer(m[1], 'base64').toString('ascii');
+    header = Buffer.from(m[1], 'base64').toString('ascii');
     let cred = header.split(':');
     let auth = {};
     let parts = cred[0].split('-');
@@ -55,25 +53,49 @@ const parse = header=>{
     return auth;
 };
 
-const calc = (opt, make_short)=>{
+const _calc = opt=>{
     const parts = ['lum'];
     for (let key in _.pick(opt, fields))
     {
-        let val = opt[key], bool = boolean_part[key];
-        if (!val || val=='*')
+        let val = opt[key];
+        if (val==='' || val=='*' || val==undefined)
             continue;
         if (key=='city')
             val = escape_geo(val);
-        if (make_short && short[key])
-            key = short[key];
         parts.push(key);
-        if (!bool)
-        {
-            parts.push((''+val).toLowerCase()
-                .replace(/ /g, make_short ? '' : '_'));
-        }
+        if (!boolean_part[key])
+            parts.push((''+val).toLowerCase().replace(/ /g, '_'));
     }
     return parts.join('-');
 };
 
-module.exports = {parse, calc, fields};
+const calculate_username = function(opt){
+    if (opt.ext_proxy)
+    {
+        return Object.assign({password: opt.password},
+            _.pick(opt.ext_proxy, 'username', 'password'));
+    }
+    const opt_usr = Object.assign({}, opt);
+    if (opt_usr.debug && opt_usr.debug==lpm_config.server_default.debug)
+        delete opt_usr.debug;
+    if (opt_usr.route_err &&
+        opt_usr.route_err==lpm_config.server_default.route_err)
+    {
+        delete opt_usr.route_err;
+    }
+    if (opt_usr.ip)
+        delete opt_usr.session;
+    if (!opt_usr.mobile)
+        delete opt_usr.mobile;
+    if (opt_usr.mobile && opt_usr.os)
+        opt_usr.mobile = false;
+    else if (opt_usr.os)
+        delete opt_usr.mobile;
+    return {
+        username: _calc(opt_usr),
+        password: opt.password,
+    };
+};
+
+
+module.exports = {parse, calculate_username};
