@@ -6,18 +6,23 @@ import setdb from '../../util/setdb.js';
 import React from 'react';
 import $ from 'jquery';
 import classnames from 'classnames';
-import {Modal, Loader, Textarea, Tooltip, Warnings} from './common.js';
-import {ga_event, presets} from './util.js';
-import Pure_component from '../../www/util/pub/pure_component.js';
+import {Loader, Warnings, Code, Preset_description,
+    with_www_api} from './common.js';
+import {Nav_tabs, Nav_tab} from './common/nav_tabs.js';
+import {report_exception} from './util.js';
+import presets from './common/presets.js';
+import Pure_component from '/www/util/pub/pure_component.js';
 import {withRouter} from 'react-router-dom';
+import prism from 'prismjs';
+import instructions from './instructions.js';
+import Tooltip from './common/tooltip.js';
+import {Textarea, Select_zone} from './common/controls.js';
+import Zone_description from './common/zone_desc.js';
+import {Modal} from './common/modals.js';
+import {T} from './common/i18n.js';
+import './css/proxy_add.less';
 
 const Proxy_add = withRouter(class Proxy_add extends Pure_component {
-    presets_opt = Object.keys(presets).map(p=>{
-        let key = presets[p].title;
-        if (presets[p].default)
-            key = `Default (${key})`;
-        return {key, value: p};
-    });
     state = {
         zone: '',
         preset: 'session_long',
@@ -25,57 +30,50 @@ const Proxy_add = withRouter(class Proxy_add extends Pure_component {
         cur_tab: 'proxy_lum',
         error_list: [],
     };
-    componentWillMount(){
-        this.setdb_on('head.settings', settings=>{
-            if (!settings)
+    static getDerivedStateFromProps(props, state){
+        const is_unblocker = ()=>{
+            if (!state.zones)
+                return false;
+            const zone_name = state.zone || state.zones.def;
+            const zone = state.zones.zones.find(p=>p.name==zone_name) || {};
+            const plan = zone.plan || {};
+            return plan.type=='unblocker';
+        };
+        return {preset: is_unblocker() ? 'unblocker' :
+            state.preset=='unblocker' ? 'session_long' : state.preset};
+    }
+    componentDidMount(){
+        this.setdb_on('head.proxies_running', proxies_running=>{
+            if (!proxies_running)
                 return;
-            this.setdb_on('head.consts', consts=>{
-                if (consts)
-                    this.prepare(consts, settings);
-            });
+            this.setState({proxies_running});
+        });
+        this.setdb_on('head.zones', zones=>{
+            if (!zones)
+                return;
+            this.setState({zones});
         });
     }
-    // XXX krzysztof: clean up zones and plans logic; use /api/zones
-    prepare = (consts, settings)=>{
-        const zones = (consts.proxy.zone.values||[]).filter(z=>{
-            const plan = z.plans && z.plans.slice(-1)[0] || {};
-            return !plan.archive && !plan.disable;
-        });
-        let def;
-        if (zones[0] && !zones[0].value && (def = settings.zone||zones[0].key))
-            zones[0] = {key: `Default (${def})`, value: ''};
-        this.setState({consts, zones, def});
-    };
     persist = ()=>{
-        const preset = this.state.preset;
-        let form;
+        const form = {};
         if (this.state.cur_tab=='proxy_lum')
         {
-            form = {
-                last_preset_applied: preset,
-                zone: this.state.zone||this.state.def,
-                proxy_type: 'persist',
-            };
-            const zone = this.state.zones.filter(z=>z.key==form.zone)[0]||{};
-            form.password = zone.password;
-            presets[preset].set(form);
+            form.preset = this.state.preset;
+            form.zone = this.state.zone;
         }
         else
         {
-            form = {
-                proxy_type: 'persist',
-                ext_proxies: this.state.parsed_ips_list,
-                // XXX krzysztof: move it to backend session: true
-                session: true,
-            };
+            form.preset = 'rotating';
+            form.ext_proxies = this.state.parsed_ips_list;
         }
+        form.proxy_type = 'persist';
+        presets.get(form.preset).set(form);
         const _this = this;
         return etask(function*(){
-            this.on('uncaught', e=>{
-                console.log(e);
-                ga_event('add-new-port', 'failed saved', e.message);
+            this.on('uncaught', e=>_this.etask(function*(){
+                yield report_exception(e, 'proxy_add.Proxy_add.persist');
                 _this.setState({show_loader: false});
-            });
+            }));
             const proxies = yield ajax.json({url: '/api/proxies_running'});
             let port = 24000;
             proxies.forEach(p=>{
@@ -83,7 +81,6 @@ const Proxy_add = withRouter(class Proxy_add extends Pure_component {
                     port = p.port+1;
             });
             form.port = port;
-            // XXX krzysztof: switch fetch->ajax
             const raw_resp = yield window.fetch('/api/proxies', {
                 method: 'POST',
                 headers: {'Content-Type': 'application/json'},
@@ -92,14 +89,15 @@ const Proxy_add = withRouter(class Proxy_add extends Pure_component {
             const resp = yield raw_resp.json();
             if (resp.errors)
                 return resp;
-            ga_event('add-new-port', 'successfully saved');
             return {port};
         });
     };
-    save(opt={}){
+    save = (opt={})=>{
         const _this = this;
         this.etask(function*(){
-            this.on('uncaught', e=>{ console.log(e); });
+            this.on('uncaught', e=>_this.etask(function*(){
+                yield report_exception(e, 'proxy_add.Proxy_add.save');
+            }));
             _this.setState({show_loader: true});
             const resp = yield _this.persist();
             if (resp.errors)
@@ -107,9 +105,12 @@ const Proxy_add = withRouter(class Proxy_add extends Pure_component {
                 _this.setState({error_list: resp.errors});
                 $('#add_proxy_errors').modal('show');
             }
-            $('#add_new_proxy_modal').modal('hide');
-            yield etask.sleep(500);
-            _this.setState({show_loader: false});
+            if (resp.port)
+                _this.setState({show_loader: false});
+            if (resp.port&&!_this.state.proxies_running.length)
+                _this.setState({created_port: resp.port});
+            else
+                $('#add_new_proxy_modal').modal('hide');
             if (!resp.errors)
             {
                 const proxies = yield ajax.json({url: '/api/proxies_running'});
@@ -118,6 +119,8 @@ const Proxy_add = withRouter(class Proxy_add extends Pure_component {
                     resp.port);
                 if (opt.redirect)
                 {
+                    yield etask.sleep(500);
+                    $('#add_new_proxy_modal').modal('hide');
                     const state_opt = {};
                     if (opt.field)
                         state_opt.field = opt.field;
@@ -126,62 +129,45 @@ const Proxy_add = withRouter(class Proxy_add extends Pure_component {
                 }
             }
         });
-    }
-    rule_clicked(field){
-        ga_event('add-new-port', 'rules clicked', field);
-        this.save({redirect: true, field});
-    }
-    advanced_clicked(){
-        ga_event('add-new-port', 'click save');
-        this.save({redirect: true});
-    }
-    field_changed = id=>value=>{
-        this.setState({[id]: value});
-        if (id=='zone')
-            ga_event('add-new-port', 'zone selected', value);
-        else if (id=='preset')
-        {
-            ga_event('add-new-port', 'preset selected',
-                `${this.state.preset}_${value}`);
-        }
     };
-    change_tab(id){
-        ga_event('add-new-port', 'changed proxy type', id);
-        this.setState({cur_tab: id});
-    }
+    field_changed = id=>value=>this.setState({[id]: value});
+    set_tab = id=>this.setState({cur_tab: id});
+    on_hidden = ()=>this.setState({created_port: null});
     render(){
         const disabled = this.state.cur_tab=='proxy_ext'&&
             !this.state.valid_json;
-        const Footer_wrapper = <Footer save_clicked={this.save.bind(this)}
-          disabled={disabled}/>;
-        if (!this.state.zones)
+        const Footer_wrapper = <Footer save_clicked={this.save}
+            disabled={disabled} created_port={this.state.created_port}/>;
+        if (!this.state.proxies_running || !this.state.zones)
             return null;
-        let content;
-        if (this.state.cur_tab=='proxy_lum')
-        {
-            content = <Lum_proxy
-                  zone={this.state.zone}
-                  zones={this.state.zones}
-                  on_field_change={this.field_changed.bind(this)}
-                  preset={this.state.preset}
-                  rule_clicked={this.rule_clicked.bind(this)}
-                  advanced_clicked={this.advanced_clicked.bind(this)}
-                  presets_opt={this.presets_opt}/>;
-        }
-        else if (this.state.cur_tab=='proxy_ext')
-        {
-            content = <Ext_proxy
-                  parse_error={this.state.parse_error}
-                  ips_list={this.state.ips_list}
-                  on_field_change={this.field_changed.bind(this)}/>;
-        }
+        const anim_classes = classnames('proxy_form', 'animated', {
+            proxy_created: this.state.created_port,
+            fadeOutUp: this.state.created_port,
+        });
         return <div className="lpm">
               <Loader show={this.state.show_loader}/>
               <Modal id="add_new_proxy_modal" no_header no_close
-                footer={Footer_wrapper} className="add_proxy_modal">
-                <Nav_tabs change_tab={this.change_tab.bind(this)}
-                  cur_tab={this.state.cur_tab}/>
-                {content}
+                on_hidden={this.on_hidden} footer={Footer_wrapper}
+                className="add_proxy_modal">
+                <div className={anim_classes}>
+                  <Nav_tabs_wrapper set_tab={this.set_tab}
+                    cur_tab={this.state.cur_tab}/>
+                  {this.state.cur_tab=='proxy_lum' &&
+                    <Lum_proxy
+                      def_zone={this.state.zones.def}
+                      created_port={this.state.created_port}
+                      zone={this.state.zone}
+                      on_field_change={this.field_changed}
+                      preset={this.state.preset}/>
+                  }
+                  {this.state.cur_tab=='proxy_ext' &&
+                    <Ext_proxy
+                      parse_error={this.state.parse_error}
+                      ips_list={this.state.ips_list}
+                      on_field_change={this.field_changed}/>
+                  }
+                </div>
+                <Created_port port={this.state.created_port}/>
               </Modal>
               <Modal className="warnings_modal" id="add_proxy_errors"
                 title="Errors:" no_cancel_btn>
@@ -191,10 +177,35 @@ const Proxy_add = withRouter(class Proxy_add extends Pure_component {
     }
 });
 
+const Created_port = ({port})=>{
+    if (!port)
+        return null;
+    const to_copy = instructions.code(port).shell;
+    const code = prism.highlight(to_copy, prism.languages.clike);
+    return <div className="howto animated fadeInUp">
+          <h3 style={{marginBottom: 15, marginTop: -20}}>
+            <T>Congratulation! You just created a port</T> {port}
+          </h3>
+          <span>
+            <T>
+              You can start using the port by running the following command:
+            </T>
+          </span>
+          <div className="well instructions_well">
+            <pre>
+              <Code>
+                <div dangerouslySetInnerHTML={{__html: code}}/>
+              </Code>
+            </pre>
+          </div>
+        </div>;
+};
+
 const Ext_proxy = ({ips_list, on_field_change, parse_error})=>{
-    const json_example = '[\'1.1.1.2\', \'username:password@1.2.3.4:8888\']';
+    const json_example = '[\'1.1.1.2\', '
+        +'\'my_username:my_password@1.2.3.4:8888\']';
     const placeholder = 'List of IPs to the external proxies in the following '
-        +'format: [username:password@]ip[:port], example:\n'+json_example;
+        +'format: [username:password@]ip[:port]';
     const on_change_list = val=>{
         on_field_change('ips_list')(val);
         try {
@@ -228,94 +239,96 @@ const Ext_proxy = ({ips_list, on_field_change, parse_error})=>{
         </div>;
 };
 
-const Lum_proxy = ({zone, zones, on_field_change, preset, rule_clicked,
-    presets_opt, advanced_clicked})=>
-{
-    const preset_tip = `Presets is a set of preconfigured configurations
-    for specific purposes`;
+const Lum_proxy = with_www_api(props=>{
+    const {zone, def_zone, on_field_change, preset} = props;
     const zone_tip = `Zone that will be used by this proxy port`;
-    const rule_tip = `Click to save a proxy port and move to this
-    configuration`;
     return <div className="lum_proxy">
-          <div className="fields_wrapper">
-            <div className="fields">
-              <Field icon_class="zone_icon" val={zone} options={zones}
-                title="Zone" on_change={on_field_change('zone')}
-                tooltip={zone_tip}/>
-              <Field icon_class="preset_icon" val={preset}
-                options={presets_opt} title="Preset configuration"
-                on_change={on_field_change('preset')} tooltip={preset_tip}/>
+          <div className="group">
+            <Field icon_class="zone_icon" title="Zone">
+              <Select_zone val={zone} tooltip={zone_tip}
+                on_change_wrapper={on_field_change('zone')}/>
+            </Field>
+            <Preview title={zone||def_zone}>
+              <Zone_description zone_name={zone}/>
+              <a className="link" href={`${props.www_api}/cp/zones`}
+                target="_blank" rel="noopener noreferrer"><T>Edit zone</T></a>
+            </Preview>
+          </div>
+          {preset!='unblocker' &&
+            <div className="group">
+              <Nav_tabs set_tab={on_field_change('preset')} cur_tab={preset}>
+                <Preset_nav_tab id="session_long"/>
+                <Preset_nav_tab id="rotating"/>
+              </Nav_tabs>
+              <Preview title={presets.get(preset).new_title}>
+                <Preset_description preset={preset}/>
+              </Preview>
             </div>
-          </div>
-          <div className="preview">
-            <div className="header">{presets[preset].title}</div>
-            <div className="desc">{presets[preset].subtitle}</div>
-            <ul>
-            {(presets[preset].rules||[]).map(r=>
-              <li key={r.field}>
-                <Tooltip title={rule_tip}>
-                  <a className="link" onClick={()=>rule_clicked(r.field)}>
-                    {r.label}</a>
-                </Tooltip>
-              </li>
-            )}
-            </ul>
-            <Tooltip
-              title="Creates a proxy port and moves to the configuration page">
-              <a onClick={advanced_clicked} className="link">
-                Advanced options</a>
-            </Tooltip>
-          </div>
+          }
+        </div>;
+});
+
+const Preset_nav_tab = ({id, ...props})=>{
+    const preset = presets.get(id);
+    return <Nav_tab title={preset.new_title} id={id}
+      tooltip={preset.subtitle} {...props}/>;
+};
+
+const Preview = ({title, children})=>{
+    return <div className="preview">
+          <div className="header">{title}</div>
+          {children}
         </div>;
 };
 
-const Nav_tabs = ({change_tab, cur_tab})=>
-    <div className="nav_tabs tabs">
-      <Tab on_click={change_tab} title="Luminati" id="proxy_lum"
-        cur_tab={cur_tab}
+const Nav_tabs_wrapper = ({set_tab, cur_tab})=>
+    <Nav_tabs set_tab={set_tab} cur_tab={cur_tab}>
+      <Nav_tab title="Luminati" id="proxy_lum"
         tooltip="Proxy port using your Luminati account"/>
-      <Tab on_click={change_tab} title="External" id="proxy_ext"
-        cur_tab={cur_tab}
+      <Nav_tab title="External" id="proxy_ext"
         tooltip="Proxy port configured with external IP and credentials"/>
-    </div>;
-
-const Tab = ({id, on_click, title, cur_tab, tooltip})=>{
-    const active = cur_tab==id;
-    const btn_class = classnames('btn_tab', {active});
-    return <Tooltip title={tooltip}>
-          <div onClick={()=>on_click(id)} className={btn_class}>
-            <div className={classnames('icon', id)}/>
-            <div className="title">{title}</div>
-            <div className="arrow"/>
-          </div>
-        </Tooltip>;
-};
+    </Nav_tabs>;
 
 const Field = props=>
-    <Tooltip title={props.tooltip}>
+    <T>{t=>
       <div className="field">
         <div className="field_header">
           <div className={classnames('icon', props.icon_class)}/>
-          <h4>{props.title}</h4>
+          <h4>{t(props.title)}:</h4>
         </div>
-        <select onChange={e=>props.on_change(e.target.value)}
-          value={props.val}>
-          {props.options.map((o, i)=>
-            <option key={i} value={o.value}>{o.key}</option>)}
-        </select>
+        {props.children ||
+          <Tooltip title={t(props.tooltip)}>
+            <select onChange={e=>props.on_change(e.target.value)}
+              value={props.val}>
+              {props.options.map((o, i)=>
+                <option key={i} value={o.value}>
+                  {props.i18n ? t(o.key) : o.key}
+                </option>)}
+            </select>
+          </Tooltip>
+        }
       </div>
-    </Tooltip>;
+    }</T>;
 
 const Footer = props=>{
     const save_clicked = ()=>{
-        ga_event('add-new-port', 'click advanced');
         if (props.disabled)
             return;
         props.save_clicked();
     };
+    const ok_clicked = ()=>$('#add_new_proxy_modal').modal('hide');
     const classes = classnames('btn', 'btn_lpm', 'btn_lpm_primary',
         {disabled: props.disabled});
-    return <button onClick={save_clicked} className={classes}>Save</button>;
+    return <div className="footer">
+          {!props.created_port &&
+            <button onClick={save_clicked} className={classes}>
+              <T>Save</T>
+            </button>
+          }
+          {!!props.created_port &&
+            <button onClick={ok_clicked} className={classes}><T>OK</T></button>
+          }
+        </div>;
 };
 
 export default Proxy_add;
