@@ -1,11 +1,12 @@
 #!/usr/bin/env node
 // LICENSE_CODE ZON ISC
 'use strict'; /*jslint node:true, esnext:true*/
-const Lum_common = require('./lum_common.js');
 
-class Lum_electron_index extends Lum_common {
+class Lum_electron_index {
+    constructor(argv){
+        this.argv = argv;
+    }
     run(){
-        this.init_log();
         require('./lum_electron.js').run(this.argv);
     }
 }
