@@ -1,42 +1,52 @@
 #!/usr/bin/env node
 // LICENSE_CODE ZON ISC
 'use strict'; /*jslint node:true, esnext:true*/
-
-const check_compat = ()=>{
-    delete require.cache[require.resolve('./check_compat.js')];
-    if (!require('./check_compat.js').is_env_compat())
-        process.exit();
-};
-check_compat();
-
-const _ = require('lodash');
 const etask = require('../util/etask.js');
 const zerr = require('../util/zerr.js');
+require('../lib/perr.js').run({});
+const logger = require('../lib/logger.js').child({category: 'lum_node_index'});
+const ssl = require('../lib/ssl.js');
 const lpm_config = require('../util/lpm_config.js');
-const Lum_common = require('./lum_common.js');
 const pm2 = require('pm2');
 const child_process = require('child_process');
 const path = require('path');
-const sudo_prompt = require('sudo-prompt');
+const util_lib = require('../lib/util.js');
+const upgrader = require('./upgrader.js');
 
-class Lum_node_index extends Lum_common {
+class Lum_node_index {
+    constructor(argv){
+        this.argv = argv;
+    }
+    is_daemon_running(list){
+        const daemon = list.find(p=>p.name==lpm_config.daemon_name);
+        return !!daemon &&
+            ['online', 'launching'].includes(daemon.pm2_env.status);
+    }
     pm2_cmd(command, opt){ return etask(function*pm2_cmd(){
-        this.on('uncaught', e=>zerr('PM2: Uncaught exception: '+zerr.e2s(e)));
+        this.on('uncaught', e=>{
+            if (e.message=='process name not found')
+                logger.notice('There is no running LPM daemon process');
+            else
+                logger.error('PM2: Uncaught exception: '+zerr.e2s(e));
+        });
+        this.on('finally', ()=>pm2.disconnect());
         yield etask.nfn_apply(pm2, '.connect', []);
         if (!Array.isArray(opt))
             opt = [opt];
-        yield etask.nfn_apply(pm2, '.'+command, opt);
-        yield etask.nfn_apply(pm2, '.disconnect', []);
+        return yield etask.nfn_apply(pm2, '.'+command, opt);
     }); }
-    run_daemon(){
-        let dopt = _.pick(this.argv.daemon_opt,
-            ['start', 'stop', 'delete', 'restart', 'startup']);
-        if (!Object.keys(dopt).length)
-            return;
+    start_daemon(){
+    return etask(function*start_daemon(){
+        this.on('uncaught', e=>{
+            logger.error('PM2: Uncaught exception: '+zerr.e2s(e));
+        });
+        this.on('finally', ()=>pm2.disconnect());
+        const script = path.resolve(__dirname, 'index.js');
+        logger.notice('Running in daemon: %s', script);
         const daemon_start_opt = {
-            name: this.proc_name,
-            script: this.argv.$0,
-            'merge-logs': false,
+            name: lpm_config.daemon_name,
+            script,
+            mergeLogs: false,
             output: '/dev/null',
             error: '/dev/null',
             autorestart: true,
@@ -44,115 +54,174 @@ class Lum_node_index extends Lum_common {
             restartDelay: 5000,
             args: process.argv.filter(arg=>arg!='-d'&&!arg.includes('daemon')),
         };
-        if (dopt.start)
-            this.pm2_cmd('start', daemon_start_opt);
-        else if (dopt.stop)
-            this.pm2_cmd('stop', this.proc_name);
-        else if (dopt.delete)
-            this.pm2_cmd('delete', this.proc_name);
-        else if (dopt.restart)
-            this.pm2_cmd('restart', this.proc_name);
-        else if (dopt.startup)
-        {
-            let pm2_bin = path.resolve(__dirname, '../node_modules/.bin/pm2');
-            try {
-                child_process.execSync(pm2_bin+' startup');
-                child_process.execSync(pm2_bin+' save');
-            } catch(e){
-                zerr.warn('Failed to install startup script automatically, '
-                    +`try run:\n${e.stdout.toString('utf-8')}\n${pm2_bin}`
-                    +`save`);
+        yield etask.nfn_apply(pm2, '.connect', []);
+        yield etask.nfn_apply(pm2, '.start', [daemon_start_opt]);
+        const bus = yield etask.nfn_apply(pm2, '.launchBus', []);
+        bus.on('log:out', data=>{
+            if (data.process.name!=lpm_config.daemon_name)
+                return;
+            process.stdout.write(data.data);
+            if (data.data.includes('Open admin browser') ||
+                data.data.includes('Shutdown'))
+            {
+                this.continue();
             }
-        }
-        return true;
+        });
+        yield this.wait();
+    }); }
+    stop_daemon(){
+        const _this = this;
+    return etask(function*stop_daemon(){
+        this.on('uncaught', e=>{
+            logger.error('PM2: Uncaught exception: '+zerr.e2s(e));
+        });
+        this.on('finally', ()=>pm2.disconnect());
+        yield etask.nfn_apply(pm2, '.connect', []);
+        const pm2_list = yield etask.nfn_apply(pm2, '.list', []);
+        if (!_this.is_daemon_running(pm2_list))
+            return logger.notice('There is no running LPM daemon process');
+        const bus = yield etask.nfn_apply(pm2, '.launchBus', []);
+        let start_logging;
+        bus.on('log:out', data=>{
+            if (data.process.name!=lpm_config.daemon_name)
+                return;
+            start_logging = start_logging||data.data.includes('Shutdown');
+            if (!start_logging || !data.data.includes('NOTICE'))
+                return;
+            process.stdout.write(data.data);
+        });
+        bus.on('process:event', data=>{
+            if (data.process.name==lpm_config.daemon_name &&
+                ['delete', 'stop'].includes(data.event))
+            {
+                this.continue();
+            }
+        });
+        yield etask.nfn_apply(pm2, '.stop', [lpm_config.daemon_name]);
+        yield this.wait();
+    }); }
+    restart_daemon(){
+        const _this = this;
+        return etask(function*(){
+            const pm2_list = yield _this.pm2_cmd('list');
+            if (_this.is_daemon_running(pm2_list))
+            {
+                logger.notice('Restarting daemon...');
+                yield _this.pm2_cmd('restart', lpm_config.daemon_name);
+                logger.notice('Daemon restarted');
+            }
+        });
     }
-    restart_on_child_exit(){
+    show_status(){
+        const _this = this;
+        return etask(function*status(){
+        this.on('uncaught', e=>{
+            logger.error('Status: Uncaught exception: '+zerr.e2s(e));
+        });
+        const pm2_list = yield _this.pm2_cmd('list');
+        const running_daemon = _this.is_daemon_running(pm2_list);
+        const tasks = yield util_lib.get_lpm_tasks({all_processes: true});
+        if (!tasks.length && !running_daemon)
+            return logger.notice('There is no LPM process running');
+        let msg = 'Proxy manager status:\n';
+        if (running_daemon)
+        {
+            msg += 'Running in daemon mode. You can close it by '+
+            'running \'luminati --stop-daemon\'\n';
+        }
+        msg += util_lib.get_status_tasks_msg(tasks);
+        logger.notice(msg);
+        });
+    }
+    gen_cert(){
+        logger.notice('Generating cert');
+        ssl.gen_cert();
+    }
+    restart_on_child_exit(is_upgraded){
         if (!this.child)
             return;
         this.child.removeListener('exit', this.restart_on_child_exit);
-        setTimeout(()=>this.create_child(), 5000);
+        setTimeout(()=>this.create_child(is_upgraded), 5000);
     }
-    shutdown_on_child_exit(){ process.exit(); }
-    create_child(){
-        // XXX vladislavl: setup env for debug purposes only
+    shutdown_on_child_exit(){
+        process.exit();
+    }
+    create_child(is_upgraded){
         process.env.LUM_MAIN_CHILD = true;
         this.child = child_process.fork(
             path.resolve(__dirname, 'lum_node.js'),
-            [], {stdio: 'inherit', env: process.env});
+            process.argv.slice(2), {stdio: 'inherit', env: process.env});
         this.child.on('message', this.msg_handler.bind(this));
         this.child.on('exit', this.shutdown_on_child_exit);
-        this.child.send({command: 'run', argv: this.argv});
+        this.child.send({command: 'run', argv: this.argv, is_upgraded});
     }
     msg_handler(msg){
         switch (msg.command)
         {
-        case 'shutdown_master': return process.exit();
+        case 'shutdown_master':
+            return process.exit();
         case 'restart':
             this.child.removeListener('exit', this.shutdown_on_child_exit);
-            this.child.on('exit', this.restart_on_child_exit.bind(this));
+            this.child.on('exit', this.restart_on_child_exit.bind(this,
+                msg.is_upgraded));
             this.child.kill();
             break;
         case 'upgrade':
-            this.upgrade(e=>this.child.send({command: 'upgrade_finished',
-                error: e}));
+            this.upgrade();
+            break;
+        case 'downgrade':
+            this.downgrade();
             break;
         }
     }
-    upgrade(cb){
-        const log_file = path.join(lpm_config.work_dir,
-            'luminati_upgrade.log');
-        const npm_cmd = 'npm install --unsafe-perm -g '
-            +'@luminati-io/luminati-proxy';
-        const cmd = lpm_config.is_win ? npm_cmd :
-            `bash -c "${npm_cmd} > ${log_file} 2>&1"`;
-        const opt = {name: 'Luminati Proxy Manager'};
-        zerr.notice('Upgrading proxy manager');
-        sudo_prompt.exec(cmd, opt, (e, stdout, stderr)=>{
-            if (cb)
-                cb(e);
-            if (e)
-            {
-                zerr('Error during upgrade: '+zerr.e2s(e));
-                if (!lpm_config.is_win)
-                    zerr(`Look at ${log_file} for more details`);
-                return;
-            }
-            if (stderr)
-                zerr('NPM stderr: '+stderr);
-            check_compat();
-        });
-    }
-    run(){
-        if (this.run_daemon())
-            return;
-        this.init_log();
-        if (lpm_config.is_win)
-        {
-            const readline = require('readline');
-            readline.createInterface({input: process.stdin, output:
-                process.stdout}).on('SIGINT', ()=>process.emit('SIGINT'));
-        }
+    init_traps(){
         ['SIGTERM', 'SIGINT', 'uncaughtException'].forEach(sig=>{
             process.on(sig, e=>{
-                this.child.send({
-                    command: 'shutdown',
-                    reason: sig+(e ? 'master: error = '+zerr.e2s(e) : ''),
-                    error: zerr.e2s(e),
-                });
                 setTimeout(()=>process.exit(), 5000);
             });
         });
-        if (!this.argv.upgrade)
-            return this.create_child();
-        this.upgrade(e=>{
-            if (e)
-            {
-                zerr(`Error during upgrade: ${zerr.e2s(e)}`);
-                process.exit();
-            }
-            zerr.notice('Upgrade completed successfully');
-            this.create_child();
+        if (lpm_config.is_win)
+        {
+            const readline = require('readline');
+            readline.createInterface({
+                input: process.stdin,
+                output: process.stdout,
+            }).on('SIGINT', ()=>process.emit('SIGINT'));
+        }
+    }
+    upgrade(opt={}){
+        upgrader.upgrade(error=>{
+            if (this.child)
+                this.child.send({command: 'upgrade_finished', error});
+            this.restart_daemon();
         });
+    }
+    downgrade(){
+        upgrader.downgrade(error=>{
+            if (this.child)
+                this.child.send({command: 'downgrade_finished', error});
+            this.restart_daemon();
+        });
+    }
+    run(){
+        if (this.argv.startUpgrader)
+            return upgrader.start_upgrader();
+        if (this.argv.stopUpgrader)
+            return upgrader.stop_upgrader();
+        if (this.argv.daemon_opt.start)
+            return this.start_daemon();
+        if (this.argv.daemon_opt.stop)
+            return this.stop_daemon();
+        if (this.argv.status)
+            return this.show_status();
+        if (this.argv.genCert)
+            return this.gen_cert();
+        if (this.argv.upgrade)
+            return this.upgrade();
+        if (this.argv.downgrade)
+            return this.downgrade();
+        this.init_traps();
+        this.create_child();
     }
 }
 module.exports = Lum_node_index;
