@@ -1,24 +1,22 @@
 // LICENSE_CODE ZON ISC
 'use strict'; /*jslint node:true, esnext:true*/
 const path = require('path');
-const fs = require('fs');
-const readline = require('readline');
 const etask = require('../util/etask.js');
-const csv = require('../util/csv.js');
-const vipdb_csv = path.resolve(__dirname, '../db/vipdb.csv');
-const vipdb_small_csv = path.resolve(__dirname, '../db/vipdb_small_geo.csv');
+const logger = require('./logger.js');
+const lpm_file = require('../util/lpm_file.js');
+const file = require('../util/file.js');
 const E = module.exports;
 
 let data;
 let pending;
 
-const prepare_data = csv_data=>{
+const prepare_data = locs=>{
     const locations = {};
     const asns = {};
-    csv_data.forEach(d=>{
+    locs.forEach(d=>{
         if (!d)
             return;
-        const [country, asn, region, city, mob] = d[0].split('_');
+        const [country, asn, region, city, mob] = d.split('_');
         if (!locations[country])
             locations[country] = {_mob: !!mob};
         locations[country]._mob = locations[country]._mob||!!mob;
@@ -30,18 +28,45 @@ const prepare_data = csv_data=>{
             locations[country][region][city] = {_mob: !!mob};
         locations[country][region][city]._mob =
             locations[country][region][city]._mob||!!mob;
-        asns[asn] = {_mob: !!mob};
+        if (!asns[country])
+            asns[country] = {};
+        if (!asns[country][asn])
+            asns[country][asn] = {_mob: !!mob};
+        asns[country][asn]._mob = asns[country][asn]._mob||!!mob;
     });
     return {locations, asns};
 };
 
-E.all_locations = ()=>etask(function*(){
-    yield E.ensure_data();
+E.all_locations = api_request=>etask(function*(){
+    yield E.ensure_data(api_request);
     return data;
 });
 
-E.ensure_data = ()=>etask(function*(){
-    if (data)
+const get_current_vipdb_data = api_request=>etask(function*(){
+    let res = [];
+    let main = yield api_request({
+        endpoint: '/lpm/vipdb/main',
+        no_throw: 1,
+        force: 1,
+    });
+    if (main.statusCode!=200)
+        logger.warn('Vipdb main could not be fetched');
+    else
+        res = res.concat((main.body||'').split('\n'));
+    const small = yield api_request({
+        endpoint: '/lpm/vipdb/small_geo',
+        no_throw: 1,
+        force: 1,
+    });
+    if (small.statusCode!=200)
+        logger.warn('Vipdb small could not be fetched');
+    else
+        res = res.concat((small.body||'').split('\n'));
+    return res.map(v=>v.split(';')[0]);
+});
+
+E.ensure_data = api_request=>etask(function*(){
+    if (data || data===null)
         return;
     if (pending)
     {
@@ -49,16 +74,31 @@ E.ensure_data = ()=>etask(function*(){
         return yield this.wait();
     }
     pending = [];
-    const csv_data = [];
-    readline.createInterface({input: fs.createReadStream(vipdb_csv)})
-    .on('line', l=>csv_data.push(csv.to_arr(l)[0]))
-    .on('close', this.continue_fn());
-    yield this.wait();
-    readline.createInterface({input: fs.createReadStream(vipdb_small_csv)})
-    .on('line', l=>csv_data.push(csv.to_arr(l)[0]))
-    .on('close', this.continue_fn());
-    yield this.wait();
-    data = prepare_data(csv_data);
+    const vipdb_path = path.resolve(lpm_file.work_dir, 'vipdb.csv');
+    try { data = JSON.parse(file.read_e(vipdb_path)); }
+    catch(e){
+        if (e.code!='ENOENT')
+            logger.error('Failed to parse vipdb %s', e.message);
+    }
+    let current_vipdb_version;
+    try {
+        const res = yield api_request({
+            endpoint: '/lpm/vipdb/version', force: 1});
+        current_vipdb_version = res.body;
+    } catch(e){
+        return logger.warn('Unable to get current vipdb version');
+    }
+    if (!data || data.version!=current_vipdb_version)
+    {
+        logger.notice('Fetching vipdb data. New version %s available '
+        +'(current is %s)', current_vipdb_version, data && data.version);
+        const new_vipdb = yield get_current_vipdb_data(api_request);
+        if (!new_vipdb)
+            return;
+        data = {vipdb: new_vipdb, version: current_vipdb_version};
+        file.write_e(vipdb_path, JSON.stringify(data));
+    }
+    data = prepare_data(data && data.vipdb || []);
     pending.forEach(p=>p.continue());
     pending = null;
 });
