@@ -1,24 +1,25 @@
 // LICENSE_CODE ZON ISC
 'use strict'; /*jslint node:true, esnext:true*/
-const lokijs = require('lokijs');
-const lfsa = require('lokijs/src/loki-fs-structured-adapter.js');
+const lokijs = require('lum_lokijs');
+const lfsa = require('lum_lokijs/src/loki-fs-structured-adapter.js');
 const etask = require('../util/etask.js');
 const string = require('../util/string.js');
 const file = require('../util/file.js');
 const qw = string.qw;
-const log = require('./log.js');
+const logger = require('./logger.js').child({category: 'Loki'});
 
 const E = module.exports = Loki;
 
-function Loki(path, log_lvl){
+function Loki(path){
     this.path = path;
     this.colls = {};
-    this.db_names = qw`port status_code hostname protocol`;
-    this.log = log('loki', log_lvl);
+    this.stats_db_names = qw`port status_code hostname protocol`;
+    this.db_names = [].concat(this.stats_db_names, 'request');
+    this.ready = false;
 }
 
 E.prototype.prepare = etask._fn(function*loki_prepare(_this){
-    if (file.is_file(_this.path)&&!file.size_e(_this.path))
+    if (file.is_file(_this.path) && !file.size_e(_this.path))
         file.rm_rf_e(_this.path+'*');
     const adapter = new lfsa();
     _this.loki = new lokijs(_this.path, {
@@ -34,39 +35,105 @@ E.prototype.prepare = etask._fn(function*loki_prepare(_this){
             _this.loki.addCollection(db, {unique: [db], indices: ['reqs']});
         _this.colls[db] = _this.loki.getCollection(db);
     });
+    _this.ready = true;
+    logger.debug('loki db ready');
 });
 
 E.prototype.save = etask._fn(function*loki_save(_this){
+    if (!_this.loki)
+        return;
     const et = this;
-    _this.log.notice('Saving stats to the file');
+    logger.notice('Saving stats to the file');
     _this.loki.saveDatabase(function(err){
         if (!err)
             return et.continue(true);
-        _this.log.error(err);
+        logger.error(err);
         et.continue(err);
     });
     yield this.wait();
 });
 
 E.prototype.stats_clear = function(){
-    this.db_names.forEach(db=>this.colls[db].findAndRemove());
+    this.stats_db_names.forEach(db=>this.colls[db].findAndRemove());
 };
 
-E.prototype.stats_clear_by_port = function(port){
-    this.colls.port.findAndRemove({key: port});
+E.prototype.stats_clear_by_ports = function(ports){
+    this.colls.port.findAndRemove({key: {$in: ports}});
+};
+
+E.prototype.request_trunc = function(max_logs){
+    if (!this.ready)
+        return;
+    const count = this.colls.request.count();
+    const rm = count-max_logs;
+    if (rm>0)
+        this.colls.request.chain().simplesort('timestamp').limit(rm).remove();
+};
+
+E.prototype.request_process = function(data, max_logs){
+    data.in_bw = data.in_bw||0;
+    data.out_bw = data.out_bw||0;
+    data.bw = data.in_bw+data.out_bw;
+    try {
+        this.colls.request.insert(data);
+        this.request_trunc(max_logs);
+    } catch(e){
+        logger.error('request_process: %s', e.message);
+    }
+};
+
+E.prototype.requests_count = function(query){
+    if (!Object.keys(query).length)
+        query = undefined;
+    return this.colls.request.count(query);
+};
+
+E.prototype.request_get_by_id = function(uuid){
+    return this.colls.request.findOne({uuid});
+};
+
+E.prototype.requests_get = function(query, sort, limit, skip){
+    if (!limit)
+        limit = undefined;
+    return this.colls.request.chain()
+        .find(query)
+        .simplesort(sort.field, sort.desc)
+        .offset(skip)
+        .limit(limit)
+        .data();
+};
+
+E.prototype.requests_clear = function(ports){
+    const query = ports ? {port: {$in: ports}} : undefined;
+    this.colls.request.findAndRemove(query);
+};
+
+E.prototype.requests_sum_in = function(query){
+    return this.colls.request.chain().find(query).mapReduce(
+        r=>r.in_bw, arr=>arr.reduce((acc, el)=>acc+el, 0));
+};
+
+E.prototype.requests_sum_out = function(query){
+    return this.colls.request.chain().find(query).mapReduce(
+        r=>r.out_bw, arr=>arr.reduce((acc, el)=>acc+el, 0));
 };
 
 E.prototype.stats_process = function(data){
+    logger.debug('stats_process start');
     if (!data.out_bw && !data.in_bw)
         return;
     data.in_bw = data.in_bw||0;
     data.out_bw = data.out_bw||0;
-    this.db_names.forEach(f=>{
+    this.stats_db_names.forEach(f=>{
         if (!data[f])
             return;
         const search_opt = {key: data[f]};
-        const s = this.colls[f].findOne(search_opt)||{in_bw: 0, out_bw: 0,
-            reqs: 0, success: 0};
+        const s = this.colls[f].findOne(search_opt) || {
+            in_bw: 0,
+            out_bw: 0,
+            reqs: 0,
+            success: 0,
+        };
         this.colls[f].findAndRemove(search_opt);
         this.colls[f].insert({
             key: data[f],
@@ -76,6 +143,7 @@ E.prototype.stats_process = function(data){
             success: s.success+data.success,
         });
     });
+    logger.debug('stats_process finish');
 };
 
 E.prototype.stats_get = function(){
@@ -93,10 +161,20 @@ E.prototype.stats_get = function(){
 };
 
 E.prototype.stats_group_by = function(group_by, count=20){
+    if (!this.ready)
+        return [];
     if (!count)
         count = undefined;
     try {
         return this.colls[group_by].chain().simplesort('reqs', true)
         .limit(count).data({removeMeta: true});
-    } catch(e){ return []; }
+    } catch(e){
+        logger.error('stats_group_by: %s', e.message);
+        return [];
+    }
+};
+
+E.prototype.stop = function(){
+    if (this.loki)
+        this.loki.autosaveDisable();
 };
