@@ -1,424 +1,145 @@
 // LICENSE_CODE ZON ISC
-'use strict'; /*jslint node:true, esnext:true, evil: true*/
+'use strict'; /*jslint node:true, esnext:true, evil:true*/
 const _ = require('lodash');
+const cheerio = require('cheerio');
 const zutil = require('../util/util.js');
-const date = require('../util/date.js');
+const {migrate_trigger} = require('../util/rules_util.js');
 const etask = require('../util/etask.js');
-const http = require('http');
-const decode_body = require('./util.js').decode_body;
-const assign = Object.assign, {SEC, MIN, HOUR, DAY} = date.ms;
+const zurl = require('../util/url.js');
+const request = require('request');
+const logger = require('./logger.js').child({category: 'Rules'});
+const lutil = require('./util.js');
 
-let write_http_reply = (_stream, res, headers)=>{
-    headers = assign(headers||{}, res.headers||{});
-    if (_stream.x_hola_context)
-        headers['x-hola-context'] = _stream.x_hola_context;
-    if (_stream.cred)
-        headers['x-lpm-authorization'] = _stream.cred;
-    _stream.resp_written = true;
-    if (_stream instanceof http.ServerResponse)
-        return _stream.writeHead(res.statusCode, res.statusMessage, headers);
-    let head = `HTTP/1.1 ${res.statusCode} ${res.statusMessage}\r\n`;
-    for (let field in headers)
-        head += `${field}: ${headers[field]}\r\n`;
-    _stream.write(head+'\r\n');
-};
-
-module.exports = class Rules {
-    constructor(luminati, rules){
-        rules = rules||{};
-        const priority_cmp = (a, b)=>a.priority-b.priority;
-        this.luminati = luminati;
+class Rules {
+    constructor(server, rules=[]){
+        this.server = server;
         this.rules = zutil.clone_deep(rules);
-        this._post = (this.rules.post||[]).map(p=>{
-            p.url_re = this._url_regexp(p.url);
-            p.need_body = p.res.filter(r=>r.body||r.action&&r.action.process)
-            .length>0;
-            p.priority = p.priority||10;
-            if (p.res&&p.res.some(pres=>_.get(pres, 'action.reserve_session')))
-                p.pool_key = 'reserve_session';
-            return p;
-        }).sort(priority_cmp);
-        // XXX vladislavl: temp solution for activation fast_pool
-        // must be removed when permanent solution is released
-        this._post.forEach(p=>{
-            if (!p.res||!p.res.some(r=>_.get(r, 'action.fast_pool_session')))
-                return;
-            p.pool_key = `fast_pool:${p.url}`;
-            this.rules.pre = this.rules.pre||[];
-            this.rules.pre.push({url: p.url, priority: p.priority
-                ||Number.MAX_SAFE_INTEGER, pool_key: p.pool_key});
+        this.rules = this.rules.map(rule=>{
+            if (!rule.type || !rule.trigger_code)
+                rule = migrate_trigger(rule);
+            rule.trigger = new Trigger(rule);
+            return rule;
         });
-        this._pre = (this.rules.pre||[]).map(p=>{
-            p.url_re = this._url_regexp(p.url);
-            p.priority = p.priority||10;
-            return p;
-        }).sort(priority_cmp);
-    }
-    _url_regexp(_url){
-        if (!_url||_url=='*'||_url=='**')
-            return new RegExp('');
-        let r;
-        try { r = new RegExp(_url); }
-        catch(e){ r = new RegExp('$a'); }
-        return r;
-    }
-    get_time(t){
-        let n = t.match(/^(\d+)(ms|sec|min|hr|day)?$/);
-        if (!n)
-        {
-            t = t+'ms';
-            n = t.match(/^(\d+)(ms|sec|min|hr|day)?$/);
-        }
-        if (!n)
+        const rule_to_priority = rule=>{
+            const action = rule.action||{};
+            if (action.retry)
+                return 1;
+            if (action.retry_port)
+                return 2;
             return 0;
-        t = +n[1];
-        switch (n[2])
-        {
-        case 'day': t *= DAY; break;
-        case 'hr': t *= HOUR; break;
-        case 'min': t *= MIN; break;
-        case 'sec': t *= SEC; break;
-        case 'ms': break;
-        }
-        return t;
+        };
+        this.rules.sort(function(a, b){
+            return rule_to_priority(a)-rule_to_priority(b);
+        });
     }
     pre(req, res, head){
-        const _this = this;
-    return etask(function*rules_pre(){
-        const {ctx} = req;
-        req.ctx.pool_key = req.ctx.h_reserved ? 'reserve_session' : null;
-        if (ctx.h_context=='STATUS CHECK'||!_this._pre)
+        if ('STATUS CHECK'==req.ctx.h_context)
             return;
         const _url = req.ctx.url||req.url_full||req.url;
-        for (let i=0; i<_this._pre.length; i++)
+        const opt = {url: _url};
+        for (let i=0; i<this.rules.length; i++)
         {
-            let p = _this._pre[i];
-            if (p.connect)
-            {
-                if (req.method!='CONNECT'||p.connect!=_url)
-                    continue;
-            }
-            else if (!p.url_re.test(_url))
+            const rule = this.rules[i];
+            if (rule.type!='before_send' && rule.type!='timeout')
                 continue;
-            if (p.session)
-                req.session = _this.gen_session();
-            else if (p.pool_key && !req.ctx.pool_key)
-                req.ctx.pool_key = p.pool_key;
-            // XXX krzysztof: legacy and does not work
-            if (p.timeout)
-                req.timeout = _this.get_time(p.timeout);
-            if (p.trigger_type=='min_req_time')
+            if (!rule.trigger.test(opt))
+                continue;
+            if (rule.type=='timeout' && opt.timeout)
             {
+                const _this = this;
                 req.min_req_task = etask(function*min_req_time(){
-                    yield etask.sleep(p.min_req_time);
-                    _this._action(req, res, p);
+                    yield etask.sleep(opt.timeout);
+                    _this.action(req, res, head, rule, opt);
                 });
                 return false;
             }
-            else if (p.action=='null_response')
-            {
-                _this.luminati._send_rule_mail(p.email,
-                    {type: 'URL', value: p.url}, 'Null response', _url);
-                ctx.log.debug('requested url %s matches null_response', _url);
-                return _this.send_null_response(req, res);
-            }
-            else if (p.action=='bypass_proxy')
-            {
-                _this.luminati._send_rule_mail(p.email,
-                    {type: 'URL', value: p.url}, 'Bypass proxy', _url);
-                ctx.log.debug('requested url %s matches bypass proxy', _url);
-                const resp = yield ctx.lum.send_bypass_req(req, res, head);
-                // XXX maximk: sometimes retval is not returned directly, why?
-                if (resp && resp.child && resp.child.retval)
-                    return resp.child.retval;
-                return resp;
-            }
-            else if (p.action=='direct')
-            {
-                _this.luminati._send_rule_mail(p.email,
-                    {type: 'URL', value: p.url}, 'Direct super proxy', _url);
-                ctx.log.debug('requested url %s matches direct request', _url);
-                ctx.is_direct = true;
-            }
-            else if (p.action=='switch_port')
-            {
-                let serv;
-                serv = _this.luminati.get_other_port(p.port);
-                if (!serv)
-                    return _this.send_null_response(req, res);
-                ctx.rules = serv.rules;
-                ctx.port = serv.port;
-                serv._request(req, res, head);
-                return 'switched';
-            }
+            return this.action(req, res, head, rule, opt);
         }
-    }); }
-    send_null_response(req, res){
-        const ctx = req.ctx;
-        ctx.log.debug('Returning null response: %s %s', req.method, ctx.url);
-        let status = req.method=='CONNECT' ? 501 : 200;
-        write_http_reply(res, {statusCode: status, statusMessage: 'NULL'});
-        res.end();
-        ctx.timeline.track('end');
-        ctx.response.status_code = status;
-        ctx.response.status_message = 'NULL';
-        return ctx.response;
     }
-    _cmp(rule, value){
-        if (!rule)
-            return false;
-        let type = rule.type||'==';
-        switch (type)
+    post(req, res, head, _res){
+        if ('STATUS CHECK'==req.ctx.h_context)
+            return;
+        for (let i=0; i<this.rules.length; i++)
         {
-        case '==': return rule.arg==value;
-        case '!=': return rule.arg!=value;
-        case '=~':
-            if (!rule.arg_re)
-                rule.arg_re = new RegExp(rule.arg);
-            return rule.arg_re.test(value);
-        case '!~':
-            if (!rule.arg_re)
-                rule.arg_re = new RegExp(rule.arg);
-            return !rule.arg_re.test(value);
-        case 'in': return rule.arg.includes(value);
-        case '!in': return !rule.arg.includes(value);
-        }
-        return false;
-    }
-    cmp(rule, value){
-        if (!rule)
-            return false;
-        if (!(rule instanceof Array))
-        {
-            if (rule.name)
-                value = value[rule.name];
-            return this._cmp(rule, value);
-        }
-        for (let i=0; i<rule.length; i++)
-        {
-            let r = rule[i], v = value;
-            if (r.name)
-                v = value[rule.name];
-            if (this._cmp(r, v))
+            const rule = this.rules[i];
+            const status = _res.statusCode || _res.status_code;
+            const time_created = _.get(req, 'ctx.timeline.req.create')||0;
+            const time_passed = Date.now()-time_created;
+            const opt = {url: req.ctx.url, status, time_passed, _res};
+            if (rule.type!='after_hdr')
+                continue;
+            if (!rule.trigger.test(opt) || req.ctx.skip_rule(rule))
+                continue;
+            this.check_req_time_range(req, rule);
+            if (this.action(req, res, head, rule, opt))
                 return true;
-        }
-        return false;
-    }
-    // XXX krzysztof: probably legacy and does not work
-    post_timeout(req, res, head){
-        if (req.ctx.h_context=='STATUS CHECK'||!this._post)
-            return;
-        const _url = req.url_full||req.url;
-        for (let i=0; i<this._post.length; i++)
-        {
-            let p = this._post[i];
-            if (p.connect)
-            {
-                if (req.method!='CONNECT'||p.connect!=_url)
-                    continue;
-            }
-            else if (!p.url_re.test(_url))
-                continue;
-            for (let j=0; j<p.res.length; j++)
-            {
-                let r = p.res[j];
-                if (!r.timeout)
-                    continue;
-                if (this.action(req, res, head, null, r.action||p.action, p))
-                    return true;
-            }
-        }
-    }
-    post(req, res, head, _res, hdrs_only){
-        const ctx = req.ctx;
-        if (ctx.h_context=='STATUS CHECK'||!this._post)
-            return;
-        const _url = ctx.url;
-        for (let i=0; i<this._post.length; i++)
-        {
-            let p = this._post[i];
-            if (p.connect)
-            {
-                if (req.method!='CONNECT'||p.connect!=_url)
-                    continue;
-            }
-            else if (!p.url_re.test(_url))
-                continue;
-            for (let j=0; j<p.res.length; j++)
-            {
-                const r = p.res[j], pr_action = r.action||p.action;
-                if (hdrs_only && !r.head)
-                    continue;
-                if (this.check_req_time_range(req, r)&&_res)
-                {
-                    if (this.action(req, res, head, _res, pr_action, p,
-                        {type: r.trigger_type.replace(/_/g, ' '),
-                            value: r[r.trigger_type]}))
-                    {
-                        return true;
-                    }
-                }
-                if (r.ipban&&_res)
-                {
-                    let tl = _res.hola_headers&&
-                        _res.hola_headers['x-hola-timeline-debug']||
-                        _res.headers&&_res.headers['x-hola-timeline-debug'];
-                    if (tl)
-                    {
-                        const ip = tl.split(' ')[3];
-                        if (this.luminati.banlist.has(ip)
-                            && this.action(req, res, head, _res, pr_action, p,
-                            {type: 'IP was banned', value: ip}))
-                        {
-                            return true;
-                        }
-                    }
-                }
-                if (this.cmp(r.status, _res.statusCode)
-                    || this.cmp(r.header, _res.headers))
-                {
-                    if (this.action(req, res, head, _res, pr_action, p,
-                        {type: 'status code', value: r.status.arg}))
-                    {
-                        return true;
-                    }
-                }
-            }
         }
     }
     post_body(req, res, head, _res, body){
-        const ctx = req.ctx;
-        if (ctx.h_context=='STATUS CHECK'||!this._post)
+        if ('STATUS CHECK'==req.ctx.h_context)
             return;
-        const _body = decode_body(body, _res.headers['content-encoding']);
-        const _url = ctx.url;
-        for (let i=0; i<this._post.length; i++)
+        let _body;
+        try {
+            _body = lutil.decode_body(body, _res.headers['content-encoding']);
+        } catch(e){
+            logger.error('error decoding body: %s when requesting url %s',
+                e.message, req.ctx.url);
+            return;
+        }
+        for (let i=0; i<this.rules.length; i++)
         {
-            let p = this._post[i];
-            if (!p.need_body || !p.url_re.test(_url))
+            const rule = this.rules[i];
+            const status = _res.statusCode||_res.status_code;
+            const time_passed = Date.now()-req.ctx.timeline.req.create;
+            const opt = {url: req.ctx.url, status, body: _body, time_passed,
+                _res};
+            if (rule.type!='after_body')
                 continue;
-            for (let j=0; j<p.res.length; j++)
-            {
-                const r = p.res[j], pr_action = r.action||p.action;
-                if (this.cmp(r.status, _res.statusCode))
-                {
-                    if (this.action(req, res, head, _res, pr_action, p,
-                        {type: 'status with body', value: r.status}))
-                    {
-                        return true;
-                    }
-                }
-                if (r.body && this.cmp(r.body, _body))
-                {
-                    if (this.action(req, res, head, _res, pr_action, p,
-                        {type: 'HTML body element', value: r.body.arg}))
-                    {
-                        return true;
-                    }
-                }
-                if (r.action&&r.action.process)
-                {
-                    this.process_response(req, _res, r.action.process, _body,
-                        r.action, {type: 'URL', value: p.url});
-                    return false;
-                }
-            }
+            if (!rule.trigger.test(opt) || req.ctx.skip_rule(rule))
+                continue;
+            this.check_req_time_range(req, rule);
+            let res_action;
+            if (res_action = this.action(req, res, head, rule, opt))
+                return res_action;
         }
     }
     post_need_body(req){
-        if (!this._post)
-            return;
-        const ctx = req.ctx;
-        const _url = ctx.url;
-        for (let i=0; i<this._post.length; i++)
-        {
-            let p = this._post[i];
-            if (!p.url_re.test(_url))
-                continue;
-            if (!p.need_body)
-                continue;
-            return true;
-        }
-        return false;
+        return this.rules.some(rule=>rule.type=='after_body');
     }
-    check_req_time_range(req, r){
-        if (!r.max_req_time && !r.min_req_time)
+    check_req_time_range(req, rule){
+        if (!rule.max_req_time && !rule.min_req_time)
             return false;
-        const ctx = req.ctx, pk = ctx.pool_key;
-        let req_max = r.max_req_time ? this.get_time(r.max_req_time)
-            : +Infinity;
-        let req_min = r.min_req_time ? this.get_time(r.min_req_time) : 0;
-        let req_time = Date.now()-ctx.timeline.req.create;
-        // XXX vladislavl: temp solution while rules are inconsistent:
-        // one logic rule in current arch - several independent srtuctures
-        // this code MUST be moved inside rule structure
-        const res = req_time<=req_max && req_time>=req_min;
-        if (!res && /fast_pool/.test(pk))
-        {
-            this.luminati.session_mgr.remove_session_from_pool(
-                ctx.session, pk);
-        }
+        const time_passed = Date.now()-req.ctx.timeline.req.create;
+        const max_time = rule.max_req_time||+Infinity;
+        const min_time = rule.min_req_time||0;
+        const res = time_passed<=max_time && time_passed>=min_time;
         return res;
     }
-    retry(req, res, head, port){
+    retry(req, res, head, opt={}){
+        const {port} = opt;
         if (!req.retry)
             req.retry = 0;
         req.retry++;
-        req.ctx.log.info('req retry %s %s', req.retry, req.ctx.url);
-        let serv;
-        if (port)
-            serv = this.luminati.get_other_port(port);
-        if (!serv)
-            serv = this.luminati;
-        req.ctx.rules = serv.rules;
-        req.ctx.port = serv.port;
-        req.ctx.proxies.forEach(p=>this.luminati._abort_proxy_req(req, p));
-        if (this.luminati.port!=port&&this.luminati.opt.handle_abort)
-            this.luminati.opt.handle_abort(req.ctx.uuid);
-        serv._request(req, res, head);
+        logger.info('req retry %s %s', req.retry, req.ctx.url);
+        (req.ctx.proxies||[]).forEach(p=>this.server.abort_proxy_req(req, p));
+        this.server.emit('retry', {port: port||this.server.port, req, res,
+            head, post: _res=>this.post(req, res, head, _res)});
         return true;
     }
-    // XXX krzysztof: WIP _can_retry is a newer implementation of can_retry
-    _can_retry(req, res, rule){
-        if (rule.action!='retry'&&rule.action!='retry_port')
-            return false;
-        const retry = req.retry||0;
-        const port = rule.retry_port;
-        if (port&&!this.luminati.get_other_port(port))
-            return false;
-        return retry<5;
+    can_retry(req, action={}){
+        const retried = req.retry||0;
+        let retry = Number(action.retry)||0;
+        if (action.retry_port)
+            retry = 20;
+        return retried<retry;
     }
-    can_retry(req, response, action){
-        let retry = req.retry||0;
-        let ret = (!action||!action.refresh_ip)&&retry<5||retry==0;
-        let port;
-        if (action)
-        {
-            let action_retry = parseInt(action.retry);
-            if (action.retry===false)
-                action_retry = 0;
-            else if (action.retry===true)
-                action_retry = 5;
-            else if (isNaN(action_retry))
-                action_retry = 0;
-            ret = !(!action.retry||action_retry<=retry||!ret);
-            port = action.retry_port;
-            if (port&&!this.luminati.get_other_port(port))
-                ret = false;
-        }
-        return ret;
-    }
-    gen_session(){
-        return 'rand'+Math.floor(Math.random()*9999999+1000000);
-    }
-    process_response(req, _res, process_rules, body, action, trigger){
-        if (action.email)
-        {
-            this.luminati.sp.spawn(this.luminati._send_rule_mail(
-                action.email, trigger, 'process data', req.url));
-        }
+    process_response(req, _res, body, rule){
+        const process_rules = rule.action.process;
         _res.headers['content-type'] = 'application/json; charset=utf-8';
         delete _res.headers['content-encoding'];
+        /* eslint-disable no-unused-vars */
+        const $ = cheerio.load(body, {xml: {normalizeWhitespace: true}});
+        /* eslint-enable no-unused-vars */
         const processed = {};
         for (let key in process_rules)
         {
@@ -432,69 +153,137 @@ module.exports = class Rules {
         }
         try {
             const new_body = JSON.stringify(processed);
-            req.ctx.response.body = [Buffer(new_body)];
+            req.ctx.response.body = [Buffer.from(new_body)];
         } catch(e){
             const new_body = JSON.stringify({
                 error: 'processing data',
                 message: e.message,
                 context: process_rules,
             });
-            req.ctx.response.body = [Buffer(new_body)];
+            req.ctx.response.body = [Buffer.from(new_body)];
         }
     }
-    // XXX krzysztof: WIP _action is a newer implementation of action
-    _action(req, res, rule){
-        if (!this._can_retry(req, res, rule))
+    send_request(req, opt, ip){
+        if (!zurl.is_valid_url(opt.url))
+            return;
+        opt.method = opt.method || 'GET';
+        const has_payload = opt.method!='GET' && !!opt.payload;
+        const url = /^https?:\/\//.test(opt.url) ? opt.url : 'http://'+opt.url;
+        const req_opt = {url, method: opt.method};
+        if (has_payload)
+        {
+            const payload = Object.assign({}, opt.payload);
+            const k = Object.keys(payload).find(_k=>payload[_k]=='$IP');
+            if (k)
+                payload[k] = ip;
+            req_opt.body = JSON.stringify(payload);
+            req_opt.headers = {
+                'Content-Type': 'application/json',
+                'Content-Length': Buffer.byteLength(req_opt.body),
+            };
+        }
+        request(req_opt, e=>{
+            if (e)
+            {
+                logger.error('error when requesting url %s : %s', req_opt.url,
+                    e.message);
+            }
+        });
+    }
+    action(req, res, head, rule, opt){
+        // XXX krzysztof: temp work-around to get IP. Should be moved to
+        // server.js and check why SMTP requests don't get to handler
+        if (req.ctx && req.ctx.session)
+        {
+            req.ctx.session.last_res = req.ctx.session.last_res||{};
+            req.ctx.session.last_res.ip = opt._res && opt._res.headers &&
+                opt._res.headers['x-luminati-ip'];
+        }
+        // XXX krzysztof: bad design - code should not be aware of tests
+        if (req.ctx && req.ctx.rule_executed)
+            req.ctx.rule_executed(rule);
+        // XXX krzysztof: find out why this.lumina.sp is sometimes null here
+        if (this.server.sp)
+            this.server.sp.spawn(this.server.send_email(rule, opt.url));
+        if (rule.action.null_response)
+            return this.server.router.send_null_response(req, res);
+        if (rule.action.bypass_proxy)
+        {
+            req.ctx.is_bypass_proxy = true;
+            return;
+        }
+        if (rule.action.direct)
+        {
+            req.ctx.is_direct = true;
+            return;
+        }
+        if (rule.action.reserve_session)
+            this.server.session_mgr.add_to_pool(req.ctx.session);
+        const _res = opt._res;
+        const headers = _res && _res.hola_headers || _res && _res.headers;
+        const ip = headers && headers['x-luminati-ip'];
+        if (rule.action.ban_ip!=undefined && _res)
+        {
+            const t = rule.action.ban_ip || 0;
+            if (ip)
+                this.server.banip(ip, t, req.ctx.session);
+        }
+        if (rule.action.ban_ip_global!=undefined && _res)
+        {
+            const t = rule.action.ban_ip_global || 0;
+            if (ip)
+                this.server.emit('banip_global', {ip, t});
+            if (req.ctx.session)
+                this.server.session_mgr.replace_session(req.ctx.session);
+        }
+        if (rule.action.ban_ip_domain!=undefined && _res)
+        {
+            const t = rule.action.ban_ip_domain || 0;
+            const domain = lutil.url2domain(opt.url);
+            if (ip)
+                this.server.banip(ip, t, req.ctx.session, domain);
+        }
+        if (rule.action.process)
+        {
+            this.process_response(req, opt._res, opt.body, rule);
             return false;
-        req.session = this.gen_session();
-        this.retry(req, res, null, rule.retry_port);
-        return true;
-    }
-    action(req, res, head, _res, action, rule, trigger){
-        const ctx = req.ctx;
-        if (action.email)
-        {
-            this.luminati.sp.spawn(this.luminati._send_rule_mail(
-                action.email, trigger, rule.res[0].action_type, req.url));
         }
-        if (action.reserve_session)
+        if (rule.action.request_url)
         {
-            this.luminati.session_mgr.add_reserve_pool_session(ctx.session,
-                rule.pool_key);
-        }
-        if (action.fast_pool_session)
-        {
-            this.luminati.session_mgr.add_fast_pool_session(ctx.session,
-                rule.pool_key, action.fast_pool_size);
-        }
-        if (!this.can_retry(req, res, action))
+            this.send_request(req, rule.action.request_url, ip);
             return false;
-        const tl = _res&&_res.hola_headers&&
-            _res.hola_headers['x-hola-timeline-debug']||
-            _res.headers&&_res.headers['x-hola-timeline-debug'];
-        if (action.ban_ip&&_res)
-        {
-            let t = this.get_time(action.ban_ip)||1;
-            if (tl)
-                this.luminati.banlist.add(tl.split(' ')[3], t);
-            req.session = this.gen_session();
         }
-        else if (action.refresh_ip&&_res&&tl)
+        if (!this.can_retry(req, rule.action))
+            return false;
+        if (rule.action.refresh_ip && opt._res && ip)
         {
-            const refresh_task = this.luminati.refresh_ip(ctx,
-                tl.split(' ')[3]);
-            this.luminati.refresh_task = refresh_task;
+            const refresh_task = this.server.refresh_ip(req.ctx, ip);
+            this.server.refresh_task = refresh_task;
         }
-        else if (action.url&&_res)
-        {
-            let _url = action.url;
-            if (_url=='location')
-                _url = _res.headers.location;
-            req.url = _url;
-        }
-        else
-            req.session = this.gen_session();
-        this.retry(req, res, head, action.retry_port);
-        return true;
+        this.server.session_mgr.refresh_sessions();
+        this.retry(req, res, head, {port: rule.action.retry_port});
+        return 'switched';
     }
-};
+}
+
+const E = module.exports = Rules;
+
+class Trigger {
+    constructor(r){
+        const code = r.trigger_code+'\nreturn trigger;';
+        try { this.func = new Function(code)(); }
+        catch(e){ logger.warn('trigger: invalid function'); }
+    }
+    test(opt){
+        if (!this.func)
+        {
+            logger.warn('trigger: invalid function');
+            return false;
+        }
+        try { return this.func(opt||{}); }
+        catch(e){ logger.warn('trigger: function failed'); }
+        return false;
+    }
+}
+
+E.t = {Trigger};
