@@ -1,25 +1,25 @@
 // LICENSE_CODE ZON ISC
 'use strict'; /*jslint node:true, esnext:true*/
-const socks = require('@luminati-io/socksv5');
+const socks = require('lum_socksv5');
 const net = require('net');
 const lutil = require('./util.js');
 const util = require('util');
 const events = require('events');
-const log = require('./log.js');
+const logger = require('./logger.js').child({category: 'SOCKS'});
 const etask = require('../util/etask.js');
 const zerr = require('../util/zerr.js');
 module.exports = Socks;
 
 function Socks(opt){
     events.EventEmitter.call(this);
-    this._log = log(`socks:${opt.port}-${opt.port}`, opt.log);
     this.opt = opt;
     this.server = socks.createServer(this.handler.bind(this));
     this.server.useAuth(socks.auth.None());
     this.connections = {};
     this.next_id = 0;
     this.server.on('error', err=>{
-        this._log.error('server error %O', err);
+        if (err.message)
+            err.message = 'SOCKS '+err.message;
         this.emit('error', err);
     });
 }
@@ -27,13 +27,16 @@ function Socks(opt){
 util.inherits(Socks, events.EventEmitter);
 
 Socks.prototype.handler = function(info, accept, deny){
+    let ip = info.srcAddr;
+    if (!this.opt.serv.is_whitelisted_ip(ip))
+    {
+        logger.warn('access denied: %s is not whitelisted', ip);
+        this.opt.serv.emit('access_denied', ip);
+        return deny();
+    }
     if (info.dstPort==80)
         return this.proxy_http(info, accept, deny);
-    // XXX krzysztof: this is quick fix. solution should check zon's config
-    // to see what ports are allowed
-    if (info.dstPort==443||info.dstPort==43||info.dstPort==25)
-        return this.proxy_https(info, accept, deny);
-    this.bypass(info, accept, deny);
+    return this.proxy_https(info, accept, deny);
 };
 
 Socks.prototype.get_src_header = function(info){
@@ -46,38 +49,36 @@ Socks.prototype.lpm_connect = function(){
 
 Socks.prototype.proxy_https = function(info, accept, deny){
     const dst = this.lpm_connect();
-    this._log.debug('Socks https connection %O', info);
     dst.on('connect', ()=>{
         dst.write(
             `CONNECT ${info.dstAddr}:${info.dstPort} HTTP/1.1\r\n`+
-            `Host: ${info.dstAddr}:${info.dstPort}\r\n`+
-            `${this.get_src_header(info)}\r\n\r\n`);
-
+            `Host: ${info.dstAddr}:${info.dstPort}\r\n\r\n`);
         dst.once('data', d=>{
             let res;
             try {
                 res = lutil.parse_http_res(d.toString('utf8'));
             } catch(e){
-                this._log.error('could not parse http resp %s', zerr.e2s(e));
+                logger.error('could not parse http resp %s', zerr.e2s(e));
                 dst.end();
                 deny();
                 return;
             }
             if (res.headers['x-lpm-error'])
             {
-                this._log.error('connection error %s:%s %s %s: %s',
+                logger.warn('connection error %s:%s %s %s: %s',
                     info.dstAddr, info.dstPort, res.status_code,
                     res.status_message, res.headers['x-lpm-error']||'');
             }
-            let socket = accept(true);
+            const socket = accept(true);
             if (!socket)
             {
-                this._log.warn('connection closed by client %O', info);
+                logger.warn('connection closed by client %s:%s', info.dstAddr,
+                    info.dstPort);
                 dst.end();
                 return;
             }
             socket.pipe(dst).pipe(socket);
-            this._handle_connection(socket, dst);
+            this.handle_connection(info, socket, dst);
         });
     });
 };
@@ -85,39 +86,39 @@ Socks.prototype.proxy_https = function(info, accept, deny){
 Socks.prototype.proxy_http = function(info, accept, deny){
     const dst = this.lpm_connect();
     const socket = accept(true);
-    this._log.debug('Socks http connection %O', info);
-    socket.once('data', data=>{
-        let d = data.toString().split('\r\n');
-        d.splice(d.indexOf(''), 0, this.get_src_header(info));
-        dst.write(d.join('\r\n'));
-        socket.pipe(dst).pipe(socket);
-    });
-    return this._handle_connection(socket, dst);
-};
-
-Socks.prototype.bypass = function(info, accept, deny){
-    const socket = accept(true);
-    this._log.debug('Socks direct connection %O', info);
-    let dst = net.connect(info.dstPort, info.dstAddr);
+    if (!socket)
+    {
+        logger.warn('connection closed by client %s:%s', info.dstAddr,
+            info.dstPort);
+        dst.end();
+        return;
+    }
     socket.pipe(dst).pipe(socket);
-    this._handle_connection(socket, dst);
+    return this.handle_connection(info, socket, dst);
 };
 
-Socks.prototype._handle_connection = function(socket, dst){
+Socks.prototype.get_remote_ip = function(port){
+    const connection = Object.values(this.connections)
+        .find(c=>c.dst.localPort==port);
+    return connection && connection.remote_ip;
+};
+
+Socks.prototype.handle_connection = function(info, socket, dst){
     const id = this.next_id++;
     socket._connectionId = dst._connectionId = id;
-    this.connections[id] = {socket, dst};
-    const end = src=>()=>{
-        this._log.debug('connection end %s connid:%s', src, id);
-        delete this.connections[id];
+    this.connections[id] = {socket, dst, remote_ip: socket.remoteAddress};
+    const close = ()=>{
+        const c = this.connections[id];
+        if (c && c.socket.destroyed && c.dst.destroyed)
+            delete this.connections[id];
     };
-    socket.on('end', end('socket'));
-    dst.on('end', end('dst'));
-    socket.on('error', error=>{
-        this._log.error('Socks connection error, %O', error);
+    socket.once('close', close);
+    dst.once('close', close);
+    socket.once('error', error=>{
+        logger.warn('Socks connection error, %s', error.message);
     });
-    dst.on('error', error=>{
-        this._log.error('Socks forward connection error, %O', error);
+    dst.once('error', error=>{
+        logger.warn('Socks forward connection error, %s', error.message);
     });
 };
 
@@ -131,21 +132,18 @@ function*set_tcp_server(_this, tcp_server){
     _this.server._srv.close(function(){ sp.continue(); });
     yield sp;
     _this.server._srv = tcp_server;
-    _this.server._srv.on('error', err=>_this.server.emit('error', err))
+    _this.server._srv.on('error', err=>{
+        _this.server.emit('error', err);
+    })
     .on('listening', ()=>_this.server.emit('listening'))
     .on('close', ()=>_this.server.emit('close'));
 });
 
-Socks.prototype.stop = function(force){
-    this._log.debug('Socks stopping');
-    if (force)
+Socks.prototype.stop = function(){
+    for (let k in this.connections)
     {
-        for (let k in this.connections)
-        {
-            let c = this.connections[k];
-            c.socket.destroy();
-            c.dst.destroy();
-        }
+        const c = this.connections[k];
+        c.socket.destroy();
+        c.dst.destroy();
     }
-    this._log.debug('Socks stopped');
 };
