@@ -28,7 +28,7 @@ class Timeline {
         return metric1-metric2;
     }
     get(name, idx, nofb){
-        if (typeof idx != 'number')
+        if (typeof idx!='number')
             idx = this.req_chain.length-1;
         if (!Array.isArray(name))
             name = [name];
@@ -45,14 +45,6 @@ class Timeline {
             this.req.end = now;
         this.req = {create: now, port};
         this.req_chain.push(this.req);
-    }
-    refresh_ip(ip){
-        this.refresh_req = {create: Date.now(), ip, port: 'Refreshing IP '+ip};
-        this.req_chain.push(this.refresh_req);
-    }
-    finish_refresh_ip(ip){
-        if (this.refresh_req)
-            this.refresh_req.end = Date.now();
     }
     toString(){
         let parts = [`timeline: t:${this.get_delta('end', true)}ms`];
