@@ -1,21 +1,23 @@
 // LICENSE_CODE ZON ISC
 'use strict'; /*jslint react:true, es6:true*/
 import React from 'react';
-import Pure_component from '../../www/util/pub/pure_component.js';
-import * as Common from './common.js';
+import Pure_component from '/www/util/pub/pure_component.js';
 import classnames from 'classnames';
 import _ from 'lodash';
+import Tooltip from './common/tooltip.js';
+import {Checkbox} from './common.js';
+import './css/chrome_widgets.less';
+import {T} from './common/i18n.js';
 
-export const Toolbar_button = ({id, tooltip, on_click, active, href})=>
-    <Tooltip title={tooltip} placement={'bottom'}>
+export const Toolbar_button = ({id, tooltip, active, href, placement,
+    ...props})=>
+    <Tooltip title={tooltip} placement={placement||'bottom'}>
       <a className={classnames('toolbar_item toolbar_button', id, {active})}
-        onClick={on_click||(()=>null)} href={href}>
-        <span className={id}/>
+        onClick={props.on_click||(()=>null)} href={href}>
+        <span className={classnames(id, 't_button', 'chrome_icon')}/>
+        {props.children}
       </a>
     </Tooltip>;
-
-export const Tooltip = props=>
-    <Common.Tooltip className="har_tooltip" {...props}/>;
 
 export const Devider = ()=><div className="devider"/>;
 
@@ -25,7 +27,8 @@ export const with_resizable_cols = (cols, Table)=>{
         cols = _.cloneDeep(cols);
         min_width = 22;
         moving_col = null;
-        style = {position: 'relative', display: 'flex', flex: 'auto'};
+        style = {position: 'relative', display: 'flex', flex: 'auto',
+            width: '100%'};
         componentDidMount(){
             this.resize_columns();
             window.document.addEventListener('mousemove', this.on_mouse_move);
@@ -113,7 +116,8 @@ export const with_resizable_cols = (cols, Table)=>{
             this.setState({moving: false});
         };
         render(){
-            return <div style={this.style} ref={this.set_ref}
+            const style = Object.assign({}, this.style, this.props.style||{});
+            return <div style={style} ref={this.set_ref}
                   className={classnames({moving: this.state.moving})}>
                   <Table {...this.props} cols={this.state.cols}
                     resize_columns={this.resize_columns}
@@ -148,3 +152,171 @@ export const Sort_icon = ({show, dir})=>{
         sort_desc: dir==1});
     return <div className="sort_icon"><span className={classes}/></div>;
 };
+
+
+import {AutoSizer, Table, Column} from 'react-virtualized';
+export class Infinite_chrome_table extends Pure_component {
+    state = {};
+    cell_renderer = function Cell_renderer(props){
+        if (props.rowData=='filler')
+            return <div className="chrome_td"></div>;
+        return <span key={props.rowData}>{props.rowData}</span>;
+    };
+    select_renderer = function Select_renderer(props){
+        if (props.rowData=='filler')
+            return <div className="chrome_td"></div>;
+        return <Checkbox
+          checked={this.props.selected_list.includes(props.rowData)}
+          on_change={()=>null}/>;
+    };
+    toggle_all = ()=>{
+        if (this.props.selected_all)
+            this.props.unselect_all();
+        else
+            this.props.select_all();
+    };
+    render(){
+        const rows = this.props.rows||[];
+        const {class_name, toolbar} = this.props;
+        return <div className="chrome">
+            <div className="main_panel vbox">
+              <Toolbar_container>
+                <Toolbar_row>
+                  <div className="title_wrapper">{this.props.title}</div>
+                </Toolbar_row>
+                {toolbar && <Toolbar_row>{toolbar}</Toolbar_row>}
+              </Toolbar_container>
+              <React.Fragment>
+                <div className={classnames('chrome_table vbox', class_name)}>
+                  <div className="tables_container header_container hack">
+                    <div className="chrome_table">
+                      <AutoSizer>
+                        {({height, width})=>
+                          <Table width={width}
+                            height={height}
+                            onRowClick={this.props.toggle}
+                            onHeaderClick={({dataKey})=>dataKey=='select' &&
+                              this.toggle_all()}
+                            gridClassName="chrome_grid"
+                            headerHeight={27}
+                            headerClassName="chrome_th"
+                            rowClassName="chrome_tr"
+                            rowHeight={22}
+                            rowCount={rows.length+1}
+                            rowGetter={({index})=>rows[index]||'filler'}>
+                            <Column key="select"
+                              cellRenderer={this.select_renderer.bind(this)}
+                              label={
+                                <Checkbox checked={this.props.selected_all}
+                                  on_change={()=>null}/>
+                              }
+                              dataKey="select"
+                              className="chrome_td"
+                              flexGrow={0}
+                              flexShrink={1}
+                              width={20}/>
+                            {this.props.cols.map(col=>
+                              <Column key={col.id}
+                                cellRenderer={this.cell_renderer}
+                                label={col.title}
+                                className="chrome_td"
+                                dataKey={col.id}
+                                flexGrow={1}
+                                width={100}/>
+                            )}
+                          </Table>
+                        }
+                      </AutoSizer>
+                    </div>
+                  </div>
+                </div>
+              </React.Fragment>
+            </div>
+          </div>;
+    }
+}
+
+export class Chrome_table extends Pure_component {
+    render(){
+        const {selectable, cols, title, children, class_name,
+            selected_all, toggle_all} = this.props;
+        const classes = classnames('chrome', 'chrome_table', class_name);
+        return <div className={classes}>
+              <div className="main_panel vbox">
+                <Toolbar_container>
+                  <Toolbar_row>
+                    <div className="title_wrapper">{title}</div>
+                  </Toolbar_row>
+                </Toolbar_container>
+                <div className="tables_container vbox">
+                  <Header_container selectable={selectable}
+                    selected_all={selected_all} toggle_all={toggle_all}
+                    cols={cols}/>
+                  <Data_container selectable={selectable} cols={cols}>
+                    {children}
+                  </Data_container>
+                </div>
+              </div>
+            </div>;
+    }
+}
+
+const Data_container = ({cols=[], children, selectable})=>{
+    return <div className="data_container">
+          <table>
+            <colgroup>
+              {selectable &&
+                <col key="select_all" style={{width: 20}}/>
+              }
+              {cols.map((c, idx)=>
+                <col key={idx} style={{width: c.width}}/>
+              )}
+            </colgroup>
+            <tbody>
+              {children}
+              <tr className="filler">
+                {selectable && <td key="select_filler"/>}
+                {cols.map(c=><td key={c.id}></td>)}
+              </tr>
+            </tbody>
+          </table>
+        </div>;
+};
+
+const Header_container = ({cols, selectable, selected_all, toggle_all})=>
+    <div className="header_container">
+      <table>
+        <colgroup>
+          {selectable && <col key="select_all" style={{width: 20}}/>}
+          {(cols||[]).map((c, idx)=><col key={idx} style={{width: c.width}}/>)}
+        </colgroup>
+        <tbody>
+          <tr>
+            {selectable &&
+              <th key="select_all" onClick={toggle_all}>
+                <Checkbox checked={selected_all} readonly/>
+              </th>
+            }
+            {(cols||[]).map((c, idx)=><th key={idx}><T>{c.title}</T></th>)}
+          </tr>
+        </tbody>
+      </table>
+    </div>;
+
+export const Toolbar_container = ({children})=>
+    <div className="toolbar_container">
+      {children}
+    </div>;
+
+export const Toolbar_row = ({children})=>
+    <div className="toolbar">
+      {children}
+    </div>;
+
+export const Search_box = ({val, on_change})=>
+    <div className="search_box">
+      <input value={val}
+        onChange={on_change}
+        type="text"
+        placeholder="Filter"/>
+    </div>;
