@@ -1,21 +1,32 @@
 // LICENSE_CODE ZON ISC
 'use strict'; /*jslint react:true, es6:true*/
 import React from 'react';
-import Pure_component from '../../www/util/pub/pure_component.js';
+import {withRouter} from 'react-router-dom';
+import Pure_component from '/www/util/pub/pure_component.js';
 import {Typeahead} from 'react-bootstrap-typeahead';
 import ajax from '../../util/ajax.js';
+import etask from '../../util/etask.js';
 import setdb from '../../util/setdb.js';
 import zurl from '../../util/url.js';
-import {Loader, Logo} from './common.js';
-import {withRouter} from 'react-router-dom';
+import {Loader, Logo, with_www_api} from './common.js';
+import {T} from './common/i18n.js';
+import './css/login.less';
+
+const token_err_msg = {
+    'bad token': 'Invalid token',
+    'token expired': 'Expired token. A new token has been sent',
+    default: 'Something went wrong'
+};
 
 const Login = withRouter(class Login extends Pure_component {
-    constructor(props){
-        super(props);
-        this.state = {password: '', username: '', loading: false};
-    }
+    state = {password: '', username: '', loading: false, two_step: false,
+        two_step_verified: false, customer_selected: false,
+        sending_email: false};
     componentDidMount(){
-        this.setdb_on('head.settings', settings=>this.setState({settings}));
+        this.setdb_on('head.argv', argv=>{
+            if (argv)
+                this.setState({argv});
+        });
         this.setdb_on('head.ver_node', ver_node=>this.setState({ver_node}));
         const url_o = zurl.parse(document.location.href);
         const qs_o = zurl.qs_parse((url_o.search||'').substr(1));
@@ -25,10 +36,15 @@ const Login = withRouter(class Login extends Pure_component {
             this.save_user();
         }
     }
-    update_password({target: {value}}){ this.setState({password: value}); }
-    update_username({target: {value}}){ this.setState({username: value}); }
-    select_customer(customer){ this.setState({customer}); }
-    save_user(){
+    update_password = ({target: {value}})=>this.setState({password: value});
+    update_username = ({target: {value}})=>this.setState({username: value});
+    select_customer = customer=>this.setState({customer});
+    select_final_customer = ()=>{
+        if (this.state.customer)
+            this.setState({customer_selected: true});
+        this.save_user();
+    };
+    save_user = ()=>{
         const creds = {};
         if (this.token)
             creds.token = this.token;
@@ -49,72 +65,134 @@ const Login = withRouter(class Login extends Pure_component {
                     update.error_message = 'Something went wrong';
                 _this.setState(update);
             });
+            this.on('finally', ()=>_this.setState({loading: false}));
             _this.setState({loading: true});
             const res = yield ajax.json({url: '/api/creds_user',
                 method: 'POST', data: creds, timeout: 60000});
             if (res.error)
             {
-                _this.setState({error_message:
-                    res.error.message||'Something went wrong'});
+                _this.setState({
+                    customer: null,
+                    user_customers: null,
+                    customer_selected: false,
+                    error_message: res.error.message||'Something went wrong',
+                });
             }
-            else if (res.customers)
+            else if (res.customers || res.ask_two_step)
             {
                 _this.setState({
                     user_customers: res.customers,
                     error_message: '',
-                    user_data: {customer: res.customers[0]},
+                    two_step: res.ask_two_step,
                 });
             }
             else
-                _this.get_in();
-            _this.setState({loading: false});
+                yield _this.get_in();
         });
-    }
-    get_in(){
+    };
+    verify_two_step = data=>{
         const _this = this;
         this.etask(function*(){
-            this.on('uncaught', _this.get_in);
-            const settings = yield ajax.json({url: '/api/settings'});
-            window.ga('set', 'dimension1', settings.customer);
-            setdb.set('head.settings', settings);
-            const consts = yield ajax.json({url: '/api/consts'});
-            setdb.set('head.consts', consts);
-            const zones = yield ajax.json({url: '/api/zones'});
-            setdb.set('head.zones', zones);
+            _this.setState({loading: true});
+            const res = yield ajax.json({url: '/api/verify_two_step',
+                method: 'POST', data, no_throw: true});
+            if (res && res.error)
+            {
+                _this.setState({loading: false, error_message:
+                    token_err_msg[res.message]||token_err_msg.default});
+            }
+            else
+                _this.setState({two_step_verified: true}, _this.save_user);
+        });
+    };
+    send_two_step_email = ()=>{
+        const _this = this;
+        this.etask(function*(){
+            this.on('finally', ()=>_this.setState({sending_email: false}));
+            _this.setState({sending_email: true});
+            const res = yield ajax.json({url: '/api/send_two_step_email',
+                method: 'POST', no_throw: true});
+            if (res && res.error)
+            {
+                _this.setState({error_message:
+                    res.error.message||'Something went wrong'});
+            }
+        });
+    };
+    get_in = ()=>{
+        const _this = this;
+        return this.etask(function*(){
+            this.on('uncaught', e=>{
+                _this.setState({error_message: 'Cannot log in: '+e.message});
+            });
+            const ets = [];
+            ets.push(etask(function*_get_settings(){
+                const settings = yield ajax.json({url: '/api/settings'});
+                setdb.set('head.settings', settings);
+            }));
+            ets.push(etask(function*_get_consts(){
+                const consts = yield ajax.json({url: '/api/consts'});
+                setdb.set('head.consts', consts);
+            }));
+            const curr_locations = setdb.get('head.locations');
+            if (!curr_locations || !curr_locations.shared_countries)
+            {
+                ets.push(etask(function*_get_locations(){
+                    const locations = yield ajax.json(
+                        {url: '/api/all_locations'});
+                    locations.countries_by_code = locations.countries.reduce(
+                    (acc, e)=>({...acc, [e.country_id]: e.country_name}), {});
+                    setdb.set('head.locations', locations);
+                }));
+            }
+            ets.push(etask(function*_get_zones(){
+                const zones = yield ajax.json({url: '/api/zones'});
+                setdb.set('head.zones', zones);
+            }));
+            ets.push(etask(function*_get_proxies_running(){
+                const proxies = yield ajax.json({url: '/api/proxies_running'});
+                setdb.set('head.proxies_running', proxies);
+            }));
+            yield etask.all(ets);
             _this.props.history.push('/overview');
         });
-    }
+    };
     render(){
         return <div className="lum_login">
               <Logo/>
               <Messages error_message={this.state.error_message}
-                settings={this.state.settings}
+                argv={this.state.argv}
                 ver_node={this.state.ver_node}/>
               <Loader show={this.state.loading}/>
               <Header/>
-              <Form
-                save_user={this.save_user.bind(this)}
+              <Form save_user={this.save_user}
                 user_customers={this.state.user_customers}
+                customer_selected={this.state.customer_selected}
+                two_step={this.state.two_step}
                 password={this.state.password}
                 username={this.state.username}
-                update_password={this.update_password.bind(this)}
-                update_username={this.update_username.bind(this)}
-                select_customer={this.select_customer.bind(this)}/>
+                loading={this.state.loading}
+                sending_email={this.state.sending_email}
+                update_password={this.update_password}
+                update_username={this.update_username}
+                select_customer={this.select_customer}
+                select_final_customer={this.select_final_customer}
+                send_two_step_email={this.send_two_step_email}
+                verify_two_step={this.verify_two_step}/>
             </div>;
     }
 });
 
-const parse_arguments = (settings={argv: ''})=>
-    settings.argv.replace(/(--password )(.+?)( --|$)/, '$1|||$2|||$3')
-    .split('|||');
+const parse_arguments = argv=>
+    argv.replace(/(--password )(.+?)( --|$)/, '$1|||$2|||$3').split('|||');
 
-const Messages = ({error_message, settings, ver_node})=>
+const Messages = ({error_message, argv, ver_node})=>
     <div>
-      {settings && settings.argv &&
+      {argv &&
         <div className="warning">
           <div className="warning_icon"/>
           The application is running with the following arguments:
-          {parse_arguments(settings).map(a=><strong key={a}>{a}</strong>)}
+          {parse_arguments(argv).map(a=><strong key={a}>{a}</strong>)}
         </div>
       }
       {error_message &&
@@ -132,9 +210,10 @@ const Node_message = ({ver_node})=>{
           <div className="warning_icon"/>
           <div>
             <div>
-              The recommended version of node.js is
-              <strong> {ver_node.recommended}</strong>. You are using version
-              <strong> {ver_node.current&&ver_node.current.raw}</strong>.
+              <span>The recommended version of node.js is </span>
+              <strong>{ver_node.recommended}</strong>.
+              <span>You are using version </span>
+              <strong>{ver_node.current && ver_node.current.raw}</strong>.
             </div>
             <div>
               Please upgrade your node using nvm, nave or visit
@@ -155,12 +234,10 @@ const Node_message = ({ver_node})=>{
 
 const Header = ()=>
     <div className="login_header">
-      <h3>Login with your Luminati account</h3>
+      <h3><T>Login with your Luminati account</T></h3>
     </div>;
 
-const Form = ({user_customers, save_user, update_password, update_username,
-    select_customer, password, username})=>
-{
+const Form = props=>{
     const google_login_url = 'https://accounts.google.com/o/oauth2/v2/auth?'
     +'client_id=943425271003-8ibddns3o1ftp59t2su8c3psocph9v1d.apps.'
     +'googleusercontent.com&response_type=code&redirect_uri='
@@ -172,59 +249,104 @@ const Form = ({user_customers, save_user, update_password, update_username,
             l.protocol+'//'+l.hostname+':'+(l.port||80)+'?api_version=3');
         window.location = href;
     };
-    if (user_customers)
+    if (props.user_customers && !props.customer_selected)
     {
-        return <Customers_form
-              user_customers={user_customers}
-              save_user={save_user}
-              select_customer={select_customer}/>;
+        return <Customers_form user_customers={props.user_customers}
+              select_final_customer={props.select_final_customer}
+              select_customer={props.select_customer}/>;
     }
-    return <First_form
-          password={password}
-          username={username}
+    if (props.two_step)
+    {
+        return <Two_step_form verify_two_step={props.verify_two_step}
+              send_two_step_email={props.send_two_step_email}
+              email={props.username} verifying_token={props.loading}
+              sending_email={props.sending_email}/>;
+    }
+    return <First_form password={props.password}
+          username={props.username}
           google_click={google_click}
-          save_user={save_user}
-          update_password={update_password}
-          update_username={update_username}/>;
+          save_user={props.save_user}
+          update_password={props.update_password}
+          update_username={props.update_username}/>;
 };
 
-const Typeahead_wrapper = ({data, disabled, on_change, val})=>
-    <Typeahead options={data} maxResults={10}
-      minLength={0} disabled={disabled} selectHintOnEnter
-      filterBy={(option, text)=>option.indexOf(text)==0}
-      onChange={on_change} selected={val}/>;
+const filter_by = (option, props)=>option.startsWith(props.text);
 
 class Customers_form extends Pure_component {
-    constructor(props){
-        super(props);
-        this.state = {};
-    }
-    on_change(e){
+    state = {cur_customer: []};
+    on_change = e=>{
         this.setState({cur_customer: e});
         this.props.select_customer(e&&e[0]);
-    }
+    };
     render(){
         return <div className="row customers_form">
               <div className="warning choose_customer">
-                Please choose a customer.</div>
+                Please choose a customer</div>
               <div className="form-group">
                 <label htmlFor="user_customer">Customer</label>
-                <Typeahead_wrapper
-                  data={this.props.user_customers}
-                  val={this.state.cur_customer}
-                  on_change={this.on_change.bind(this)}/>
+                <Typeahead id="user_customer"
+                  options={this.props.user_customers}
+                  maxResults={10}
+                  minLength={0}
+                  selectHintOnEnter
+                  filterBy={filter_by}
+                  onChange={this.on_change}
+                  selected={this.state.cur_customer}/>
               </div>
               <button
-                onClick={this.props.save_user}
+                onClick={this.props.select_final_customer}
                 className="btn btn_lpm btn_login"
                 disabled={this.props.saving_user}>
-                {this.props.saving_user ? 'Logging in...' : 'Log in'}
+                {this.props.saving_user ? 'Logging in' : 'Log in'}
               </button>
             </div>;
     }
 }
 
-class First_form extends Pure_component {
+class Two_step_form extends Pure_component {
+    state = {token: ''};
+    componentDidMount(){
+        this.props.send_two_step_email();
+    }
+    on_key_up = e=>{
+        if (e.keyCode==13)
+            this.props.verify_two_step(this.state);
+    };
+    on_token_change = e=>this.setState({token: e.target.value});
+    render(){
+        const {verify_two_step, send_two_step_email, verifying_token,
+            sending_email, email} = this.props;
+        return <T>{t=><div className="row customers_form">
+              <div className="warning choose_customer">
+                2-Step Verification
+              </div>
+              {t('A Luminati 2-Step Verification email containing a token '
+              +'was sent to ')}
+              {email}
+              {t('. The token is valid for ')+'15 '+ t('minutes.')}
+              <div className="form-group">
+                <input className="two_step_input" onKeyUp={this.on_key_up}
+                  onChange={this.on_token_change} placeholder={t('Token')}/>
+              </div>
+              {sending_email ?
+                  t('Sending email...') :
+                  <React.Fragment>
+                    {t('Can’t find it? Check your spam folder or click ')}
+                    <a className="link" onClick={send_two_step_email}>
+                      {t('here')}
+                    </a>
+                    {t(' to resend the email.')}
+                  </React.Fragment>
+              }
+              <button onClick={()=>verify_two_step(this.state)}
+                className="btn btn_lpm btn_login" disabled={verifying_token}>
+                {verifying_token ? t('Verifying...') : t('Verify')}
+              </button>
+            </div>}</T>;
+    }
+}
+
+const First_form = with_www_api(class First_form extends Pure_component {
     on_key_up = e=>{
         if (e.keyCode==13)
             this.props.save_user();
@@ -232,20 +354,20 @@ class First_form extends Pure_component {
     render(){
         const {google_click, saving_user, password, username} = this.props;
         return <div className="login_form">
-              <div>
+              <T>{t=><div>
                 <div className="row">
                   <div className="col col_google col-sm-6">
                     <div className="btn_google_wrapper">
                       <a className="btn btn_lpm btn_google"
                         onClick={google_click}>
                         <div className="img"/>
-                        Log in with Google
+                        {t('Log in with Google')}
                       </a>
                     </div>
                   </div>
                   <div className="col col_pass col-sm-6">
                     <div className="form-group">
-                      <label htmlFor="username">Email</label>
+                      <label htmlFor="username">{t('Email')}</label>
                       <input type="email"
                         name="username"
                         onChange={this.props.update_username}
@@ -253,7 +375,7 @@ class First_form extends Pure_component {
                         value={username}/>
                     </div>
                     <div className="form-group">
-                      <label htmlFor="user_password">Password</label>
+                      <label htmlFor="user_password">{t('Password')}</label>
                       <input type="password"
                         name="password"
                         onChange={this.props.update_password}
@@ -263,25 +385,25 @@ class First_form extends Pure_component {
                     <button type="submit"
                       className="btn btn_lpm btn_login"
                       onClick={this.props.save_user}>
-                      {saving_user ? 'Logging in...' : 'Log in'}
+                      {saving_user ? t('Logging in...') : t('Log in')}
                     </button>
                   </div>
                   <div className="or_circle">Or</div>
                 </div>
                 <div className="row">
                   <div className="signup">
-                    Don't have a Luminati account?
-                    <a href="https://luminati.io/?need_signup=1"
+                    {t('Don\'t have a Luminati account?')}
+                    <a href={`${this.props.www_api}/?need_signup=1`}
                       target="_blank"
                       rel="noopener noreferrer"
                       className="link">
-                      Sign up
+                      {t('Sign up')}
                     </a>
                   </div>
                 </div>
               </div>
-            </div>;
+            }</T></div>;
     }
-}
+});
 
 export default Login;
