@@ -1,46 +1,19 @@
 // LICENSE_CODE ZON
 'use strict'; /*jslint react:true, es6:true*/
 import React from 'react';
-import Pure_component from '../../../www/util/pub/pure_component.js';
-import {Netmask} from 'netmask';
+import Pure_component from '/www/util/pub/pure_component.js';
 import setdb from '../../../util/setdb.js';
 import {tabs} from './fields.js';
-import {Labeled_controller} from '../common.js';
-import {getContext} from 'recompose';
-import PropTypes from 'prop-types';
+import {Labeled_controller, Ext_tooltip} from '../common.js';
+import {withRouter} from 'react-router-dom';
 
-export const validators = {
-    number: (min, max, req=false)=>val=>{
-        val = Number(val);
-        if (isNaN(val))
-        {
-            if (req)
-                return min;
-            return undefined;
-        }
-        else if (val < min)
-            return min;
-        else if (val > max)
-            return max;
-        return val;
-    },
-    ips_list: val=>{
-        val = val.replace(/\s/g, '');
-        const ips = val.split(',');
-        const res = [];
-        ips.forEach(ip=>{
-            try { res.push(new Netmask(ip).base); }
-            catch(e){ console.log('incorrect ip format'); }
-        });
-        return res.join(',');
-    },
-};
+export const Tab_context = React.createContext('logs');
 
-export const Config = getContext({provide: PropTypes.object})(
-class Config extends Pure_component {
+export const Config = withRouter(class Config extends Pure_component {
     state = {disabled_fields: {}};
     set_field = setdb.get('head.proxy_edit.set_field');
     is_valid_field = setdb.get('head.proxy_edit.is_valid_field');
+    is_disabled_ext_proxy = setdb.get('head.proxy_edit.is_disabled_ext_proxy');
     on_blur = ({target: {value}})=>{
         if (this.props.validator)
             this.set_field(this.props.id, this.props.validator(value));
@@ -60,12 +33,13 @@ class Config extends Pure_component {
     };
     componentDidMount(){
         const val_id = this.props.val_id ? this.props.val_id : this.props.id;
-        this.setdb_on('head.proxy_edit.form.'+val_id, val=>
-            this.setState({val, show: true}));
+        this.setdb_on('head.proxy_edit.form.'+val_id, val=>{
+            this.setState({val, show: true});
+        });
         this.setdb_on('head.proxy_edit.disabled_fields', disabled_fields=>
             disabled_fields&&this.setState({disabled_fields}));
         this.setdb_on('head.defaults', defaults=>
-            defaults&&this.setState({defaults}));
+            defaults && this.setState({defaults}));
     }
     render(){
         if (!this.state.show||!this.state.defaults)
@@ -81,26 +55,37 @@ class Config extends Pure_component {
             else
                 _default = false;
         }
-        const tab_id = this.props.provide.tab_id;
+        const tab_id = this.context;
         const disabled = this.props.disabled||!this.is_valid_field(id)||
             this.state.disabled_fields[id];
+        let state;
+        let animated = false;
+        if ((state = this.props.location.state)&&state.field)
+            animated = state.field==id;
         return <Labeled_controller
               id={id}
+              animated={animated}
               sufix={this.props.sufix}
               data={this.props.data}
               type={this.props.type}
+              on_key_up={this.on_key_up}
               on_input_change={this.on_input_change}
               on_change_wrapper={this.on_change_wrapper}
               val={this.state.val===undefined ? '' : this.state.val}
               disabled={disabled}
-              min={this.props.min}
-              max={this.props.max}
+              disabled_ips={this.props.disabled_ips}
               note={this.props.note}
               placeholder={tabs[tab_id].fields[id].placeholder||''}
               on_blur={this.on_blur}
               label={tabs[tab_id].fields[id].label}
               default={_default}
               range={this.props.range}
-              tooltip={tabs[tab_id].fields[id].tooltip}/>;
+              tooltip={tabs[tab_id].fields[id].tooltip}
+              field_tooltip={disabled && this.is_disabled_ext_proxy(id) &&
+                  <Ext_tooltip/>}
+              exact={this.props.exact}
+              filter_by={this.props.filter_by}
+              no_any={this.props.no_any}/>;
     }
 });
+Config.WrappedComponent.contextType = Tab_context;
