@@ -1,12 +1,13 @@
 // LICENSE_CODE ZON ISC
 'use strict'; /*jslint react:true, es6:true*/
 import React from 'react';
+import {T} from '../common/i18n.js';
 
 export const tabs = {
     logs: {
-        fields: [],
+        fields: {},
         label: 'Logs',
-        tooltip: 'Logs of requests sent through this proxy port',
+        tooltip: 'Request logs of requests sent through this proxy port',
     },
     target: {
         label: 'Targeting',
@@ -31,7 +32,7 @@ export const tabs = {
                     <a className="link"
                       href="http://bgp.potaroo.net/cidr/autnums.html"
                       target="_blank" rel="noopener noreferrer">
-                      list
+                      <T>list</T>
                     </a>)
                     </span>,
                 tooltip: `Select specific Internet Service Provider (ISP), or
@@ -42,6 +43,10 @@ export const tabs = {
                 label: 'Carrier',
                 tooltip: 'Network provider company name',
             },
+            os: {
+                label: 'Operating System',
+                tooltip: 'Operating System',
+            },
         },
     },
     speed: {
@@ -49,46 +54,29 @@ export const tabs = {
         tooltip: 'Control the speed of your request to improve performance',
         fields: {
             proxy: {
-                label: `Super Proxy's location`,
-                tooltip: `For shorter latency to the Super Proxies you can
-                    choose the location. Please note that this does not
-                    restrict the IPs chosen country but will simply improve
-                    performances by using Super Proxies which have shorter
-                    roundtrips relative to your current location.`,
+                label: `Super Proxy location`,
+                tooltip: `Location of Super Proxy is chosen automatically.
+                    Sometimes choosing Chinese Super Proxy works better when
+                    also using Chinese peers`,
+            },
+            race_reqs: {
+                label: 'Parallel race requests',
+                tooltip: `Sends multiple requests in parallel via different
+                    Super Proxies and uses the fastest request.`,
+                placeholder: 'Number of parallel requests'
+            },
+            throttle: {
+                label: 'Throttle requests',
+                tooltip: `Throttle requests above the given number.
+                    Allow a maximal number of parallel requests.`,
+                ext: true,
             },
             dns: {
                 label: 'DNS lookup',
                 tooltip: 'Location of DNS resolve',
             },
-            pool_size: {
-                label: 'Pool size',
-                tooltip: `Maintain number of IPs that will be pinged constantly
-                    - must have keep_alive to work properly`,
-                ext: true,
-            },
-            race_reqs: {
-                label: 'Parallel race requests',
-                tooltip: `Sends multiple requests in parallel via different
-                    super proxies and uses the fastest request`,
-                placeholder: 'Number of parallel requests'
-            },
-            proxy_count: {
-                label: 'Minimum number of super proxies',
-                tooltip: `Minimum number of super proxies to use in parallel`,
-            },
-            proxy_switch: {
-                label: 'Switch super proxy on failure',
-                tooltip: `Number of failed requests (status 403, 429, 502, 503)
-                    to switch to different super proxy`,
-            },
-            throttle: {
-                label: 'Throttle requests',
-                tooltip: `Throttle requests above the given number. Allow
-                    maximum number of parallel requests`,
-                ext: true,
-            },
             reverse_lookup: {
-                label: 'Reverse resolve',
+                label: 'Reverse DNS lookup',
                 tooltip: 'Resolve DNS from IP to url',
                 ext: true,
             },
@@ -107,59 +95,48 @@ export const tabs = {
         tooltip: 'Define custom action for specific rule',
         fields: {
             trigger_type: {
-                label: 'Rule type',
-                tooltip: `In every request the response will be analyzed.
-                    if the configured Trigger rule is true, the Action
-                    will be executed automatically`,
+                label: 'Trigger',
+                tooltip: `In every request, the response will be analyzed. if
+                    the configured Trigger rule is true, then the Action will
+                    be executed automatically`,
             },
             body_regex: {
                 label: 'String to be scanned in body (Regex)',
                 placeholder: `i.e. (captcha|robot)`,
-                tooltip: `A string(regular expression) to be scanned in the
+                tooltip: `A string (regular expression) to be scanned in the
                     body of the response`,
             },
             min_req_time: {
                 label: 'Request time more than',
                 placeholder: '500',
-                tooltip: `Any request time above the given value in
-                    milliseconds will trigger the action`,
+                tooltip: `Any request time above the given value (in
+                    milliseconds) will trigger the action`,
             },
             max_req_time: {
                 label: 'Request time less than',
                 placeholder: '500',
-                tooltip: `Any request time below the given value in
-                    milliseconds will trigger the action`,
+                tooltip: `Any request time below the given value (in
+                    milliseconds) will trigger the action`,
             },
             trigger_url_regex: {
                 label: `Optional regex to apply only for specific URLs`,
                 placeholder: `i.e. example.com`,
                 tooltip: `Enable trigger only for certain URLs. You can type
-                    regex manually or generate it by clicking on formats above
-                    the input`,
+                    regex manually or generate it by choosing from the offered
+                    formats above the input.`,
             },
-            status_code: {
-                label: 'Status code string to be scanned',
+            status: {
+                label: `Status code (string or regexp)`,
                 tooltip: `Status code to be scanned in the response headers`,
             },
-            status_custom: {
-                label: 'Custom status code (regex)',
-                placeholder: `i.e. (2..|3..|404)`,
-                tooltip: `A string(regular expression) to be scanned in the
-                    head of the response`,
-            },
             action: {
-                label: 'Action type',
-                tooltip: `The action to be executed when rule is met`,
-            },
-            fast_pool_size: {
-                label: 'Fast pool size',
-                tooltip: `System will store fast IPs up to the selected pool
-                    size number. Once pool size is reached, the system will
-                    use IPs from the fast IPs pool to route requests`,
+                label: 'Action',
+                tooltip: `The action to be executed when the rule's condition
+                    is met`,
             },
             retry_number: {
-                label: 'Number of retries',
-                tooltip: 'maximum number of retries to execute',
+                label: 'Retries',
+                tooltip: 'Maximum number of retries to execute',
                 placeholder: `e.g. '5'`,
             },
             retry_port: {
@@ -168,25 +145,40 @@ export const tabs = {
             },
             switch_port: {
                 label: 'Switch port that will be used to make the request',
-                tooltip: `By using this action you can split your traffic
+                tooltip: `Using this action you can split your traffic
                     between ports`,
             },
             ban_ip_duration: {
                 label: 'Ban IP for',
-                tooltip: 'will remove the IP for a defined amount of time',
+                tooltip: `Ban the IP for a defined amount of time. Choose 0 to
+                    ban forever`,
             },
-            ban_ip_custom: {label: 'Custom duration'},
+            type: {
+                label: 'Schedule',
+                tooltip: `The timing for the execution of the triggered
+                    function. You can try to activate the rule before sending
+                    the request for a specific URL, after receiving headers
+                    (to check status code for example), after body (to check
+                    the whole response), or after a specified amount of
+                    time (to timeout the request and retry).`,
+            },
             process: {label: 'Processing rule'},
+            request_url: {label: 'URL'},
+            request_method: {label: 'Method'},
+            request_payload: {label: 'Payload', tooltip: `Payload sent in the
+                request. The IP used for the request will be sent in a field
+                if its value equals "$IP"`},
             send_email: {
                 label: 'Send email',
-                tooltip: `Every time the rule is triggered and executed the
-                    email notification will be sent to the provided address`,
+                tooltip: `Every time the rule is triggered and executed, the
+                        email notification will be sent to the provided
+                        address`,
             },
             email: {
                 label: 'Email address',
                 tooltip: `Notifications will be sent to this email address.
-                    If you want to use another email you can add it to the
-                    list in control panel by going to the link below`,
+                        If you want to use another email you can add it to the
+                        list in control panel by going to the link below`,
             },
         },
     },
@@ -194,85 +186,37 @@ export const tabs = {
         label: 'IP control',
         tooltip: 'Set the conditions for which your IPs will change',
         fields: {
-            ip: {
-                label: 'Data center IP',
-                tooltip: `Choose specific data center IP. to ensure
-                    all requests are executed using specific Data Center IP.
-                    To view the pool of your IPs take a look at 'pool size'
-                    option`,
-                placeholder: 'insert IP value from your pool',
-            },
-            vip: {
-                label: 'gIP',
-                tooltip: `Choose specific gIP to ensure all requests are
-                    executed using specific gIP. To view the pool of your gIPs
-                    take a look at 'pool size' option`,
-                placeholder: 'insert gIP id',
-            },
-            pool_type: {
-                label: 'Pool type',
-                tooltip: `How to pull the IPs - roundrobin / sequential`,
-                ext: true,
-            },
-            keep_alive: {
-                label: 'Keep-alive',
-                tooltip: `LPM will ping an IP to keep the session alive`,
-                ext: true,
-            },
-            session_random: {
-                label: 'Random session',
-                tooltip: `Each session is generated randomly. Disable it if you
-                    want to define sessions manually`,
+            pool_size: {
+                label: 'Pool size',
+                tooltip: `A pool of sessions that is used for the requests.
+                    Each session is associated with an IP. For non-static zones
+                    the sessions are pinged every 45 seconds to keep them
+                    alive`,
                 ext: true,
             },
             session: {
                 label: 'Explicit session',
-                tooltip: `Insert session ID to maintain the same ip for as long
-                    as possible.`,
-                placeholder: `e.g. session-1234`,
+                tooltip: `Insert a session ID. You cannot use dash "-"
+                    character`,
+                placeholder: `e.g. test_session`,
             },
             sticky_ip: {
-                label: 'Sticky IP',
-                tooltip: `When connecting to remote LPM server stick sessions
-                    to each computer. Each connected computer will receive
-                    a unique session`,
+                label: 'Session per machine',
+                tooltip: `Keep separate sessions / IPs for each connecting
+                    device from an external network. Useful when the LPM is
+                    hosted on a server and multiple machines send requests`,
                 ext: true,
             },
-            max_requests: {
-                label: 'Max requests',
-                tooltip: `Change session based on number of requests can be a
-                    range or a fixed number. when using browser it should be
-                    taken into consideration that one page load will attempt
-                    multiple requests under the hood`,
+            rotate_session: {
+                label: 'Rotate IPs',
+                tooltip: 'Change session/IP on each request',
                 ext: true,
             },
-            session_duration: {
-                label: 'Session duration',
-                tooltip: `Change session every fixed number of seconds`,
-                ext: true,
-            },
-            seed: {
-                label: 'Session ID seed',
-                tooltip: `A string that will be used to maintain unified
-                    session structure. Each new session will be attached with
-                    a serial number starting from 1. (e.g. session seed 'test'
-                    will have the following sessions: test-1, test-2 ..)`,
-                placeholder: `e.g. test_session`,
-            }
-        },
-    },
-    debug: {
-        label: 'Debugging',
-        tooltip: 'Improve the info you receive from the Proxy Manager',
-        fields: {
-            log: {
-                label: 'Log level',
-                tooltip: `Decide which data to show in logs`,
-                ext: true,
-            },
-            debug: {
-                label: 'Luminati request debug info',
-                tooltip: `Send debug info on every request`,
+            session_termination: {
+                label: 'Session termination',
+                tooltip: `Stop sending new requests when the peer IP becomes
+                    unavailable and redirect to confirmation page before new IP
+                    is taken`,
             },
         },
     },
@@ -286,29 +230,24 @@ export const tabs = {
                 tooltip: `Choose a User-Agent header that will be used for
                     sending requests`,
             },
-            random_user_agent: {
-                label: 'Random User-Agent',
-                tooltip: 'Set a random User-Agent header for each request',
-            },
             override_headers: {
                 label: 'Override headers',
-                tooltip: `If you use a browser or other software then requests
+                tooltip: `If you use a browser or other software, then requests
                     may already have defined a few headers for you. Enabling
-                    this option will allow for overriding those headers`,
+                    this option will allow overriding those headers`,
             },
         },
     },
     general: {
         label: 'General',
-        tooltip: `General configuration such as port number, password and
-            bypasing`,
+        tooltip: `General configuration such as 'port number'`,
         fields: {
             internal_name: {
                 label: 'Internal name',
-                tooltip: `You can choose a name for this proxy port. If the
-                    name is passed it will be shown in the dropdowns across
-                    whole LPM. It doesn't change any behavior of the proxy
-                    port`,
+                tooltip: `You can choose a name for this proxy port.
+                    If the name is passed it will be shown in the dropdowns
+                    across whole LPM. It doesn't change any behavior of the
+                    proxy port`,
                 ext: true,
             },
             port: {
@@ -317,34 +256,23 @@ export const tabs = {
                     proxy configuration`,
                 ext: true,
             },
-            socks: {
-                label: 'SOCKS 5 port',
-                tooltip: `SOCKS 5 port is the same as proxy port and is
-                    automatically created.`,
-                ext: true,
-            },
-            password: {
-                label: 'Zone password',
-                tooltip: `Zone password as it appears in your zones page in
-                    your Luminati's control panel http://luminati.io/cp/zones`,
-            },
-            whitelist_ips: {
-                label: 'Whitelist IP access',
-                tooltip: `Grant proxy access to specific IPs. only those
-                    IPs will be able to send requests to this proxy port`,
-                placeholder: `e.g. 1.1.1.1,23.23.23.23`,
-                ext: true,
-            },
             ssl: {
-                label: 'Enable SSL logs',
-                tooltip: `Enable SSL Logs in order to save HTTPs requests`,
+                label: 'SSL analyzing',
+                tooltip: `Enable SSL request logs in order to save HTTPs
+                    requests`,
                 ext: true,
             },
-            iface: {
-                label: 'Interface',
-                tooltip: 'Define a specific network interface on which '
-                    +'the local machine is running',
-                ext: true,
+            proxy_connection_type: {
+                label: 'Connection to Super Proxy',
+                tooltip: 'Connection type between LPM and Super Proxy',
+            },
+            route_err: {
+                label: 'Route',
+                tooltip: `<div><b>pass_dyn:</b> If a request can't pass via `
+                    +`peer, automatically pass it via the Super Proxy`
+                    +`<br><b>block</b> If a `
+                    +`request can't pass via peer, block it and don't `
+                    +`send via Super Proxy</div>`,
             },
             multiply: {
                 label: 'Multiply proxy port',
@@ -353,23 +281,57 @@ export const tabs = {
             },
             multiply_ips: {
                 label: 'Multiply proxy port per IP',
-                tooltip: `Create proxy port for every selected IP from the
+                tooltip: `Create a proxy port for every selected IP from the
                     pool`,
             },
             multiply_vips: {
                 label: 'Multiply proxy port per gIP',
-                tooltip: `Create proxy port for every selected gIP from pool
-                    of available gIPS in your zone`,
+                tooltip: `Create a proxy port for every selected gIP from pool
+                    of available gIPs in your zone`,
             },
-            secure_proxy: {
-                label: 'SSL to super proxy',
-                tooltip: `Encrypt requests sent to super proxy to avoid
-                    detection on DNS`,
+            multiply_users: {
+                label: 'Multiply proxy port per user',
+                tooltip: `Create a proxy port for every selected LPM user
+                    configured in control panel`,
+            },
+            iface: {
+                label: 'Interface',
+                tooltip: 'Define a specific network interface on which '
+                    +'the local machine is running',
                 ext: true,
+            },
+            insecure: {
+                label: 'Ignore certificates',
+                tooltip: `It works only when SSL analyzing is enabled. You can
+                    enable this option to ignore checking certificates of SSL
+                    connection`,
+                ext: true,
+            },
+            smtp: {
+                label: 'Proxy through SMTP',
+                tooltip: `IPs that will be randomly chosen to proxy whole
+                    traffic through SMTP protocol`,
+            },
+            whitelist_ips: {
+                label: 'Whitelisted IPs',
+                tooltip: `Default access grant for this port. Only these
+                    IPs will be able to send requests to this port by default`
+            },
+            debug: {
+                label: 'Debug',
+                tooltip: 'Luminati request debug info',
             },
         },
     },
 };
+
+Object.keys(tabs).forEach(t=>{
+    const format = s=>s.replace(/\s+/g, ' ').replace(/\n/g, ' ');
+    const tab = tabs[t];
+    tab.tooltip = format(tab.tooltip);
+    for (let f in tab.fields)
+        tab.fields[f].tooltip = format(tab.fields[f].tooltip||'');
+});
 
 export const all_fields = Object.keys(tabs)
     .map(k=>tabs[k].fields)
