@@ -14,7 +14,7 @@ var E = ajax;
 var assign = Object.assign;
 E.events = new events.EventEmitter();
 E.json = function(opt){ return ajax(assign({}, opt, {json: 1})); };
-E.abort = function(ajax){ ajax.goto('abort'); };
+E.abort = function(aj){ aj.goto('abort'); };
 // XXX arik: need test
 function ajax(opt){
     var timeout = opt.timeout||20*date.ms.SEC, slow = opt.slow||2*date.ms.SEC;
@@ -24,10 +24,10 @@ function ajax(opt){
     var method = opt.method||opt.type||'GET';
     var data_type = opt.json ? 'json' : 'text';
     var t0 = Date.now();
-    var xhr;
+    var ajopt, xhr;
     zerr.debug('ajax('+data_type+') url '+url+' retry '+retry);
     return etask([function(){
-        var ajopt = {dataType: data_type, type: method, url: url,
+        ajopt = {dataType: data_type, type: method, url: url,
             data: data, timeout: timeout, xhrFields: {}};
         if (opt.headers)
             ajopt.headers = opt.headers;
@@ -40,9 +40,9 @@ function ajax(opt){
         if (opt.onuploadprogress)
         {
             ajopt.xhr = function(){
-                var xhr = $.ajaxSettings.xhr();
-                xhr.upload.onprogress = opt.onuploadprogress;
-                return xhr;
+                var _xhr = $.ajaxSettings.xhr();
+                _xhr.upload.onprogress = opt.onuploadprogress;
+                return _xhr;
             };
         }
         if (opt.multipart)
@@ -51,36 +51,63 @@ function ajax(opt){
             ajopt.processData = false;
             delete ajopt.dataType;
         }
-        return xhr = $.ajax(ajopt);
+        if (opt.async!==undefined)
+            ajopt.async = opt.async;
+        xhr = $.ajax(ajopt);
+        var _this = this;
+        xhr.done(function(v){
+            _this.continue(v); });
+        xhr.fail(function(_xhr, status_text, err){
+            if (data_type=='json' && _xhr && _xhr.status==200 &&
+                ['', 'ok', 'OK'].includes(_xhr.responseText))
+            {
+                _this.continue(null);
+                return;
+            }
+            if (!err && data_type=='text' && _xhr && _xhr.responseText)
+                err = _xhr.responseText;
+            _this.throw(err instanceof Error ? err : new Error(''+err));
+        });
+        return this.wait();
     }, function catch$(err){
+        xhr = xhr||{};
         zerr('ajax('+data_type+') failed url '+url+' data '+
             zerr.json(data).substr(0, 200)+' status: '+xhr.status+' '+
             xhr.statusText+'\nresponseText: '+
             (xhr.responseText||'').substr(0, 200));
-        if (retry)
+        if (retry && (!opt.should_retry||opt.should_retry(err, xhr, ajopt)))
             return this.return(ajax(assign({}, opt, {retry: retry-1})));
         if (xhr.statusText=='timeout')
             E.events.emit('timeout', this);
-        var error = xhr.statusText||'no_status';
-        if (opt.ret_err_code && xhr.statusText && xhr.status)
-            error = xhr.status+' - '+error;
+        if (xhr.status==403)
+            E.events.emit('unauthorized', this);
         if (opt.no_throw)
-            return {error: error, message: xhr.responseText};
-        var err_obj = new Error(error);
-        err_obj.hola_info = {data_type: data_type, url: url,
-            status: xhr.status, response_text: xhr.responseText};
-        err_obj.x_error = xhr.getResponseHeader('X-Luminati-Error') ||
+        {
+            return {
+                err: err,
+                url: url,
+                method: method,
+                status: +xhr.status || 0,
+                data: get_res_data(xhr),
+                xhr: xhr,
+                // legacy
+                error: xhr.statusText||'no_status', message: xhr.responseText,
+            };
+        }
+        err.hola_info = {url: url, method: method, status: xhr.status,
+            data: get_res_data(xhr), response_text: xhr.responseText};
+        err.x_error = xhr.getResponseHeader('X-Luminati-Error') ||
             xhr.getResponseHeader('X-Hola-Error');
-        throw err_obj;
-    }, function(data){
+        throw err;
+    }, function(_data){
         var t = Date.now()-t0;
         zerr[t>slow ? 'err' : 'debug'](
             'ajax('+data_type+') '+(t>slow ? 'SLOW ' : 'ok ')+t+'ms url '+url);
         if (t>slow && perr)
             perr({id: 'be_ajax_slow', info: t+'ms '+url});
         if (E.do_op)
-            E.do_op(data&&data.do_op);
-        return this.return(data);
+            E.do_op(_data&&_data.do_op);
+        return this.return(_data);
     }, function abort(){
         // reachable only via E.abort
         xhr.abort();
@@ -101,5 +128,17 @@ function ajax(opt){
         return ajax(opt);
     };
 });
+
+function get_res_data(xhr){
+    if (xhr.responseJSON!=null && xhr.responseJSON!=='')
+        return xhr.responseJSON;
+    var content_type = xhr.getResponseHeader('content-type')||'';
+    if (xhr.responseText && content_type.includes('application/json'))
+    {
+        try { return JSON.parse(xhr.responseText); }
+        catch(e){ }
+    }
+    return xhr.responseText||'';
+}
 
 return E; }); }());
