@@ -18,7 +18,6 @@ E.list = {
     AS: 'American Samoa',
     AD: 'Andorra',
     AO: 'Angola',
-    AP: 'Asia/Pacific',
     AI: 'Anguilla',
     AQ: 'Antarctica',
     AG: 'Antigua And Barbuda',
@@ -40,12 +39,10 @@ E.list = {
     BM: 'Bermuda',
     BT: 'Bhutan',
     BO: 'Bolivia',
-    BQ: 'Bonaire, Sint Eustatius and Saba',
+    BQ: 'Bonaire (Caribbean Netherlands)',
     BA: 'Bosnia And Herzegovina',
     BW: 'Botswana',
-    BV: 'Bouvet Island',
     BR: 'Brazil',
-    IO: 'British Indian Ocean Territory',
     BN: 'Brunei',
     BG: 'Bulgaria',
     BF: 'Burkina Faso',
@@ -66,7 +63,7 @@ E.list = {
     CG: 'Congo',
     CK: 'Cook Islands',
     CR: 'Costa Rica',
-    CI: "Cote D'Ivorie (Ivory Coast)",
+    CI: "Cote D'Ivoire (Ivory Coast)",
     HR: 'Croatia (Hrvatska)',
     CU: 'Cuba',
     CW: 'Curacao',
@@ -77,7 +74,6 @@ E.list = {
     DJ: 'Djibouti',
     DM: 'Dominica',
     DO: 'Dominican Republic',
-    TP: 'East Timor',
     EC: 'Ecuador',
     EG: 'Egypt',
     SV: 'El Salvador',
@@ -91,8 +87,7 @@ E.list = {
     FJ: 'Fiji',
     FI: 'Finland',
     FR: 'France',
-    FX: 'France, Metropolitan',
-    GF: 'French Guinea',
+    GF: 'French Guiana',
     PF: 'French Polynesia',
     TF: 'French Southern Territories',
     GA: 'Gabon',
@@ -105,15 +100,14 @@ E.list = {
     GR: 'Greece',
     GL: 'Greenland',
     GD: 'Grenada',
-    GP: 'Guadeloupe',
     GU: 'Guam',
+    GP: 'Guadeloupe',
     GT: 'Guatemala',
     GG: 'Guernsey',
     GN: 'Guinea',
     GW: 'Guinea-Bissau',
     GY: 'Guyana',
     HT: 'Haiti',
-    HM: 'Heard And McDonald Islands',
     HN: 'Honduras',
     HK: 'Hong Kong',
     HU: 'Hungary',
@@ -196,7 +190,7 @@ E.list = {
     PT: 'Portugal',
     PR: 'Puerto Rico',
     QA: 'Qatar',
-    RE: 'Reunion',
+    RE: 'Réunion',
     RO: 'Romania',
     RU: 'Russia',
     RW: 'Rwanda',
@@ -205,10 +199,8 @@ E.list = {
     KN: 'Saint Kitts And Nevis',
     LC: 'Saint Lucia',
     MF: 'Saint Martin',
-    PM: 'Saint Pierre And Miquelon',
     VC: 'Saint Vincent And The Grenadines',
     SM: 'San Marino',
-    SX: 'Sint Maarten',
     ST: 'Sao Tome And Principe',
     SA: 'Saudi Arabia',
     SN: 'Senegal',
@@ -216,6 +208,7 @@ E.list = {
     SC: 'Seychelles',
     SL: 'Sierra Leone',
     SG: 'Singapore',
+    SX: 'Sint Maarten (Dutch part)',
     SK: 'Slovak Republic',
     SI: 'Slovenia',
     SB: 'Solomon Islands',
@@ -228,7 +221,6 @@ E.list = {
     LK: 'Sri Lanka',
     SD: 'Sudan',
     SR: 'Suriname',
-    SJ: 'Svalbard And Jan Mayen',
     SZ: 'Swaziland',
     SE: 'Sweden',
     CH: 'Switzerland',
@@ -252,7 +244,6 @@ E.list = {
     AE: 'United Arab Emirates',
     UK: 'United Kingdom',
     US: 'United States',
-    UM: 'United States Minor Outlying Islands',
     UY: 'Uruguay',
     UZ: 'Uzbekistan',
     VU: 'Vanuatu',
@@ -265,7 +256,6 @@ E.list = {
     EH: 'Western Sahara',
     WS: 'Western Samoa',
     YE: 'Yemen',
-    YU: 'Yugoslavia',
     ZM: 'Zambia',
     ZW: 'Zimbabwe',
 };
@@ -298,12 +288,9 @@ E.adj_list = {
     BM: {name: 'Bermuda', adj: 'Bermudan'},
     BT: {name: 'Bhutan', adj: 'Bhutanese'},
     BO: {name: 'Bolivia', adj: 'Bolivian'},
-    BQ: {name: 'Bonaire', adj: 'Bonaire'},
     BA: {name: 'Bosnia and Herzegovina', adj: 'Bosnian'},
     BW: {name: 'Botswana', adj: 'Botswanan'},
-    BV: {name: 'Bouvet Island', adj: 'Bouvet Island'},
     BR: {name: 'Brazil', adj: 'Brazilian'},
-    IO: {name: 'British Indian Ocean Territory', adj: 'BIOT'},
     BN: {name: 'Brunei', adj: 'Bruneian'},
     BG: {name: 'Bulgaria', adj: 'Bulgarian'},
     BF: {name: 'Burkina Faso', adj: 'Burkinabé'},
@@ -352,7 +339,6 @@ E.adj_list = {
     GR: {name: 'Greece', adj: 'Greek'},
     GL: {name: 'Greenland', adj: 'Greenlandic'},
     GD: {name: 'Grenada', adj: 'Grenadian'},
-    GP: {name: 'Guadeloupe', adj: 'Guadeloupe'},
     GU: {name: 'Guam', adj: 'Guamanian'},
     GT: {name: 'Guatemala', adj: 'Guatemalan'},
     GG: {name: 'Guernsey', adj: 'Channel Island'},
@@ -379,7 +365,6 @@ E.adj_list = {
     KZ: {name: 'Kazakhstan', adj: 'Kazakhstani'},
     KE: {name: 'Kenya', adj: 'Kenyan'},
     KI: {name: 'Kiribati', adj: 'I-Kiribati'},
-    XK: {name: 'Kosovo', adj: 'Kosovan'},
     KW: {name: 'Kuwait', adj: 'Kuwaiti'},
     KG: {name: 'Kyrgyzstan', adj: 'Kyrgyzstani'},
     LA: {name: 'Laos', adj: 'Lao, Laotian'},
@@ -437,7 +422,6 @@ E.adj_list = {
     PT: {name: 'Portugal', adj: 'Portuguese'},
     PR: {name: 'Puerto Rico', adj: 'Puerto Rican'},
     QA: {name: 'Qatar', adj: 'Qatari'},
-    RE: {name: 'Réunion', adj: 'Réunionese'},
     RO: {name: 'Romania', adj: 'Romanian'},
     RU: {name: 'Russia', adj: 'Russian'},
     RW: {name: 'Rwanda', adj: 'Rwandan'},
@@ -445,7 +429,6 @@ E.adj_list = {
     KN: {name: 'Saint Kitts and Nevis', adj: 'Kittitian'},
     LC: {name: 'Saint Lucia', adj: 'Saint Lucian'},
     MF: {name: 'Saint Martin', adj: 'Saint-Martinoise'},
-    PM: {name: 'Saint Pierre and Miquelon', adj: 'Saint-Pierrais'},
     VC: {name: 'Saint Vincent and the Grenadines', adj: 'Saint Vincentian'},
     SM: {name: 'San Marino', adj: 'Sammarinese'},
     SA: {name: 'Saudi Arabia', adj: 'Saudi'},
@@ -454,7 +437,6 @@ E.adj_list = {
     SC: {name: 'Seychelles', adj: 'Seychellois'},
     SL: {name: 'Sierra Leone', adj: 'Sierra Leonean'},
     SG: {name: 'Singapore', adj: 'Singaporean'},
-    SX: {name: 'Sint Maarten', adj: 'Sint Eustatius'},
     SK: {name: 'Slovakia', adj: 'Slovak'},
     SI: {name: 'Slovenia', adj: 'Slovenian'},
     SB: {name: 'Solomon Islands', adj: 'Solomon Island'},
@@ -497,39 +479,247 @@ E.adj_list = {
     ZW: {name: 'Zimbabwe', adj: 'Zimbabwean'},
 };
 
-// XXX vladislavl: merge with country list and fill zones for other countries
 E.timezone = {
-    AU: 'Australia/Sydney',
-    BA: 'Europe/Sarajevo',
-    BE: 'Europe/Brussels',
-    BY: 'Europe/Minsk',
-    CA: 'Canada/Eastern',
-    CN: 'Asia/Shanghai',
+    CI: 'Africa/Abidjan',
+    GH: 'Africa/Accra',
+    ET: 'Africa/Addis_Ababa',
+    DZ: 'Africa/Algiers',
+    ER: 'Africa/Asmara',
+    ML: 'Africa/Bamako',
+    CF: 'Africa/Bangui',
+    GM: 'Africa/Banjul',
+    GW: 'Africa/Bissau',
+    MW: 'Africa/Blantyre',
+    CG: 'Africa/Brazzaville',
+    BI: 'Africa/Bujumbura',
+    EG: 'Africa/Cairo',
+    MA: 'Africa/Casablanca',
+    GN: 'Africa/Conakry',
+    SN: 'Africa/Dakar',
+    TZ: 'Africa/Dar_es_Salaam',
+    DJ: 'Africa/Djibouti',
+    CM: 'Africa/Douala',
+    EH: 'Africa/El_Aaiun',
+    SL: 'Africa/Freetown',
+    BW: 'Africa/Gaborone',
+    ZW: 'Africa/Harare',
+    ZA: 'Africa/Johannesburg',
+    SS: 'Africa/Juba',
+    UG: 'Africa/Kampala',
+    SD: 'Africa/Khartoum',
+    RW: 'Africa/Kigali',
+    CD: 'Africa/Kinshasa',
+    NG: 'Africa/Lagos',
+    GA: 'Africa/Libreville',
+    TG: 'Africa/Lome',
+    AO: 'Africa/Luanda',
+    ZM: 'Africa/Lusaka',
+    GQ: 'Africa/Malabo',
+    MZ: 'Africa/Maputo',
+    LS: 'Africa/Maseru',
+    SZ: 'Africa/Mbabane',
+    SO: 'Africa/Mogadishu',
+    LR: 'Africa/Monrovia',
+    KE: 'Africa/Nairobi',
+    TD: 'Africa/Ndjamena',
+    NE: 'Africa/Niamey',
+    MR: 'Africa/Nouakchott',
+    BF: 'Africa/Ouagadougou',
+    BJ: 'Africa/Porto-Novo',
+    ST: 'Africa/Sao_Tome',
+    LY: 'Africa/Tripoli',
+    TN: 'Africa/Tunis',
+    NA: 'Africa/Windhoek',
+    AI: 'America/Anguilla',
+    AG: 'America/Antigua',
+    AR: 'America/Argentina/Buenos_Aires',
+    AW: 'America/Aruba',
+    PY: 'America/Asuncion',
+    BB: 'America/Barbados',
+    BZ: 'America/Belize',
+    CO: 'America/Bogota',
+    VE: 'America/Caracas',
+    KY: 'America/Cayman',
     CR: 'America/Costa_Rica',
-    DE: 'Europe/Berlin',
-    EE: 'Europe/Tallinn',
-    ES: 'Europe/Madrid',
-    FR: 'Europe/Paris',
-    GB: 'Europe/London',
-    HK: 'Asia/Hong_Kong',
-    IL: 'Israel',
+    CW: 'America/Curacao',
+    DM: 'America/Dominica',
+    SV: 'America/El_Salvador',
+    GL: 'America/Godthab',
+    TC: 'America/Grand_Turk',
+    GD: 'America/Grenada',
+    GT: 'America/Guatemala',
+    EC: 'America/Guayaquil',
+    GY: 'America/Guyana',
+    CU: 'America/Havana',
+    JM: 'America/Jamaica',
+    BO: 'America/La_Paz',
+    PE: 'America/Lima',
+    NI: 'America/Managua',
+    MF: 'America/Marigot',
+    MQ: 'America/Martinique',
+    MX: 'America/Mexico_City',
+    UY: 'America/Montevideo',
+    MS: 'America/Montserrat',
+    BS: 'America/Nassau',
+    BR: 'America/Sao_Paulo',
+    PA: 'America/Panama',
+    SR: 'America/Paramaribo',
+    HT: 'America/Port-au-Prince',
+    TT: 'America/Port_of_Spain',
+    PR: 'America/Puerto_Rico',
+    CL: 'America/Santiago',
+    DO: 'America/Santo_Domingo',
+    BL: 'America/St_Barthelemy',
+    KN: 'America/St_Kitts',
+    LC: 'America/St_Lucia',
+    VI: 'America/St_Thomas',
+    VC: 'America/St_Vincent',
+    HN: 'America/Tegucigalpa',
+    VG: 'America/Tortola',
+    AQ: 'Antarctica/McMurdo',
+    YE: 'Asia/Aden',
+    KZ: 'Asia/Almaty',
+    JO: 'Asia/Amman',
+    TM: 'Asia/Ashgabat',
+    IQ: 'Asia/Baghdad',
+    BH: 'Asia/Bahrain',
+    AZ: 'Asia/Baku',
+    TH: 'Asia/Bangkok',
+    LB: 'Asia/Beirut',
+    KG: 'Asia/Bishkek',
+    BN: 'Asia/Brunei',
     IN: 'Asia/Colombo',
-    KR: 'Asia/Seoul',
-    LT: 'Europe/Vilnius',
-    MD: 'Europe/Istanbul',
-    MT: 'Europe/Malta',
-    NL: 'Europe/Amsterdam',
-    NO: 'Europe/Oslo',
+    LK: 'Asia/Colombo',
+    SY: 'Asia/Damascus',
+    BD: 'Asia/Dhaka',
+    TL: 'Asia/Dili',
+    AE: 'Asia/Dubai',
+    TJ: 'Asia/Dushanbe',
+    PS: 'Asia/Gaza',
+    VN: 'Asia/Ho_Chi_Minh',
+    HK: 'Asia/Hong_Kong',
+    ID: 'Asia/Jakarta',
+    AF: 'Asia/Kabul',
+    PK: 'Asia/Karachi',
+    NP: 'Asia/Kathmandu',
+    MY: 'Asia/Kuala_Lumpur',
+    KW: 'Asia/Kuwait',
+    MO: 'Asia/Macau',
     PH: 'Asia/Manila',
-    PL: 'Poland',
-    RU: 'Europe/Moscow',
+    OM: 'Asia/Muscat',
+    CY: 'Asia/Nicosia',
+    KH: 'Asia/Phnom_Penh',
+    KP: 'Asia/Pyongyang',
+    QA: 'Asia/Qatar',
+    MM: 'Asia/Rangoon',
+    SA: 'Asia/Riyadh',
+    KR: 'Asia/Seoul',
+    CN: 'Asia/Shanghai',
     SG: 'Asia/Singapore',
-    UA: 'Europe/Kiev',
-    UK: 'Europe/London',
-    US: 'US/Eastern',
+    TW: 'Asia/Taipei',
     UZ: 'Asia/Tashkent',
+    GE: 'Asia/Tbilisi',
+    IR: 'Asia/Tehran',
+    BT: 'Asia/Thimphu',
+    JP: 'Asia/Tokyo',
+    MN: 'Asia/Ulaanbaatar',
+    LA: 'Asia/Vientiane',
+    AM: 'Asia/Yerevan',
+    BM: 'Atlantic/Bermuda',
+    CV: 'Atlantic/Cape_Verde',
+    FO: 'Atlantic/Faroe',
+    IS: 'Atlantic/Reykjavik',
+    GS: 'Atlantic/South_Georgia',
+    SH: 'Atlantic/St_Helena',
+    FK: 'Atlantic/Stanley',
+    AU: 'Australia/Sydney',
+    CA: 'America/Toronto',
+    NL: 'Europe/Amsterdam',
+    AD: 'Europe/Andorra',
+    GR: 'Europe/Athens',
+    RS: 'Europe/Belgrade',
+    DE: 'Europe/Berlin',
+    SK: 'Europe/Bratislava',
+    BE: 'Europe/Brussels',
+    RO: 'Europe/Bucharest',
+    HU: 'Europe/Budapest',
+    DK: 'Europe/Copenhagen',
+    IE: 'Europe/Dublin',
+    GI: 'Europe/Gibraltar',
+    GG: 'Europe/Guernsey',
+    FI: 'Europe/Helsinki',
+    IM: 'Europe/Isle_of_Man',
+    MD: 'Europe/Istanbul',
+    TR: 'Europe/Istanbul',
+    JE: 'Europe/Jersey',
+    UA: 'Europe/Kiev',
+    PT: 'Europe/Lisbon',
+    SI: 'Europe/Ljubljana',
+    GB: 'Europe/London',
+    UK: 'Europe/London',
+    LU: 'Europe/Luxembourg',
+    ES: 'Europe/Madrid',
+    MT: 'Europe/Malta',
+    AX: 'Europe/Mariehamn',
+    BY: 'Europe/Minsk',
+    MC: 'Europe/Monaco',
+    RU: 'Europe/Moscow',
+    NO: 'Europe/Oslo',
+    FR: 'Europe/Paris',
+    ME: 'Europe/Podgorica',
+    CZ: 'Europe/Prague',
+    LV: 'Europe/Riga',
+    IT: 'Europe/Rome',
+    SM: 'Europe/San_Marino',
+    BA: 'Europe/Sarajevo',
+    MK: 'Europe/Skopje',
+    BG: 'Europe/Sofia',
+    SE: 'Europe/Stockholm',
+    EE: 'Europe/Tallinn',
+    AL: 'Europe/Tirane',
+    LI: 'Europe/Vaduz',
+    VA: 'Europe/Vatican',
+    AT: 'Europe/Vienna',
+    LT: 'Europe/Vilnius',
+    HR: 'Europe/Zagreb',
+    CH: 'Europe/Zurich',
+    MG: 'Indian/Antananarivo',
+    CX: 'Indian/Christmas',
+    CC: 'Indian/Cocos',
+    KM: 'Indian/Comoro',
+    TF: 'Indian/Kerguelen',
+    SC: 'Indian/Mahe',
+    MV: 'Indian/Maldives',
+    MU: 'Indian/Mauritius',
+    YT: 'Indian/Mayotte',
+    IL: 'Israel',
+    WS: 'Pacific/Apia',
+    NZ: 'Pacific/Auckland',
+    FM: 'Pacific/Chuuk',
+    VU: 'Pacific/Efate',
+    TK: 'Pacific/Fakaofo',
+    FJ: 'Pacific/Fiji',
+    TV: 'Pacific/Funafuti',
+    SB: 'Pacific/Guadalcanal',
+    GU: 'Pacific/Guam',
+    MH: 'Pacific/Majuro',
+    NR: 'Pacific/Nauru',
+    NU: 'Pacific/Niue',
+    NF: 'Pacific/Norfolk',
+    NC: 'Pacific/Noumea',
+    AS: 'Pacific/Pago_Pago',
+    PW: 'Pacific/Palau',
+    PN: 'Pacific/Pitcairn',
+    PG: 'Pacific/Port_Moresby',
+    CK: 'Pacific/Rarotonga',
+    MP: 'Pacific/Saipan',
+    PF: 'Pacific/Tahiti',
+    KI: 'Pacific/Tarawa',
+    TO: 'Pacific/Tongatapu',
+    WF: 'Pacific/Wallis',
+    PL: 'Europe/Warsaw',
+    US: 'America/New_York',
 };
-
 E.regions = {
     AD: ['Southern Europe', 'EMEA'],
     AE: ['Western Asia', 'EMEA'],
@@ -540,14 +730,13 @@ E.regions = {
     AM: ['Western Asia', 'EMEA'],
     AN: ['Caribbean'],
     AO: ['Middle Africa', 'EMEA'],
-    AP: ['Asia/Pacific'],
     AQ: ['Antarctica', 'APAC'],
     AR: ['South America', 'AMER'],
     AS: ['Polynesia', 'AMER'],
     AT: ['Western Europe', 'EMEA'],
     AU: ['Australia and New Zealand', 'APAC'],
     AW: ['Caribbean', 'AMER'],
-    AX: ['Scandinavia', 'Northern Europe', 'EMEA'],
+    AX: ['Scandinavia', 'EMEA', 'Northern Europe'],
     AZ: ['Western Asia', 'EMEA'],
     BA: ['South-Eastern Europe', 'EMEA'],
     BB: ['Caribbean', 'AMER'],
@@ -566,7 +755,6 @@ E.regions = {
     BR: ['South America', 'AMER'],
     BS: ['Caribbean', 'AMER'],
     BT: ['Southern Asia', 'APAC'],
-    BV: ['Antarctica', 'AMER'],
     BW: ['Southern Africa', 'EMEA'],
     BY: ['Eastern Europe', 'EMEA'],
     BZ: ['Central America', 'AMER'],
@@ -580,18 +768,18 @@ E.regions = {
     CK: ['Polynesia', 'APAC'],
     CL: ['South America', 'AMER'],
     CM: ['Middle Africa', 'EMEA'],
-    CN: ['Eastern Asia', 'APAC'],
+    CN: ['Asia', 'China'],
     CO: ['South America', 'AMER'],
     CR: ['Central America', 'AMER'],
     CU: ['Caribbean', 'AMER'],
     CV: ['Western Africa', 'EMEA'],
-    CW: ['Caribbean', ''],
+    CW: ['Caribbean', 'AMER'],
     CX: ['Australia and New Zealand', 'APAC'],
     CY: ['South-Eastern Europe', 'EMEA'],
     CZ: ['Eastern Europe', 'EMEA'],
     DE: ['Western Europe', 'EMEA'],
     DJ: ['Eastern Africa', 'EMEA'],
-    DK: ['Scandinavia', 'Northern Europe', 'EMEA'],
+    DK: ['Scandinavia', 'EMEA', 'Northern Europe'],
     DM: ['Caribbean', 'AMER'],
     DO: ['Caribbean', 'AMER'],
     DZ: ['Northern Africa', 'EMEA'],
@@ -603,22 +791,21 @@ E.regions = {
     ES: ['Southern Europe', 'EMEA'],
     ET: ['Eastern Africa', 'EMEA'],
     EU: ['Western Europe'],
-    FI: ['Scandinavia', 'Northern Europe', 'EMEA'],
+    FI: ['Scandinavia', 'EMEA', 'Northern Europe'],
     FJ: ['Melanesia', 'APAC'],
     FK: ['South America', 'AMER'],
     FM: ['Micronesia', 'APAC'],
-    FO: ['Scandinavia', 'Northern Europe', 'AMER'],
+    FO: ['Scandinavia', 'EMEA', 'Northern Europe'],
     FR: ['Western Europe', 'EMEA'],
-    FX: ['Western Europe'],
     GA: ['Middle Africa', 'EMEA'],
     GB: ['Northern Europe', 'EMEA'],
     GD: ['Caribbean', 'AMER'],
     GE: ['Western Asia', 'EMEA'],
-    GF: ['South America', 'EMEA'],
+    GF: ['South America', 'AMER'],
     GG: ['Northern Europe', 'EMEA'],
     GH: ['Western Africa', 'EMEA'],
     GI: ['Southern Europe', 'EMEA'],
-    GL: ['Scandinavia', 'Northern America', 'EMEA'],
+    GL: ['Scandinavia', 'EMEA', 'Northern Europe'],
     GM: ['Western Africa', 'EMEA'],
     GN: ['Western Africa', 'EMEA'],
     GP: ['Caribbean', 'AMER'],
@@ -629,8 +816,7 @@ E.regions = {
     GU: ['Micronesia', 'APAC'],
     GW: ['Western Africa', 'EMEA'],
     GY: ['South America', 'EMEA'],
-    HK: ['Eastern Asia', 'APAC'],
-    HM: ['Australia and New Zealand', 'AMER'],
+    HK: ['Eastern Asia', 'China'],
     HN: ['Central America', 'AMER'],
     HR: ['South-Eastern Europe', 'EMEA'],
     HT: ['Caribbean', 'AMER'],
@@ -640,10 +826,9 @@ E.regions = {
     IL: ['Western Asia', 'EMEA'],
     IM: ['Northern Europe', 'EMEA'],
     IN: ['Southern Asia', 'APAC'],
-    IO: ['Eastern Africa', 'APAC'],
     IQ: ['Western Asia', 'EMEA'],
     IR: ['Southern Asia', 'EMEA'],
-    IS: ['Scandinavia', 'Northern Europe', 'EMEA'],
+    IS: ['Scandinavia', 'EMEA', 'Northern Europe'],
     IT: ['Southern Europe', 'EMEA'],
     JE: ['Northern Europe', 'EMEA'],
     JM: ['Caribbean', 'AMER'],
@@ -701,7 +886,7 @@ E.regions = {
     NG: ['Western Africa', 'EMEA'],
     NI: ['Central America', 'AMER'],
     NL: ['Western Europe', 'EMEA'],
-    NO: ['Scandinavia', 'Northern Europe', 'EMEA'],
+    NO: ['Scandinavia', 'EMEA', 'Northern Europe'],
     NP: ['Southern Asia', 'APAC'],
     NR: ['Micronesia', 'AMER'],
     NU: ['Polynesia', 'AMER'],
@@ -714,7 +899,6 @@ E.regions = {
     PH: ['South-Eastern Asia', 'APAC'],
     PK: ['Southern Asia', 'EMEA'],
     PL: ['Eastern Europe', 'EMEA'],
-    PM: ['Northern America', 'AMER'],
     PN: ['Polynesia', 'AMER'],
     PR: ['Caribbean', 'AMER'],
     PS: ['Western Asia', 'EMEA'],
@@ -722,7 +906,7 @@ E.regions = {
     PW: ['Micronesia', 'APAC'],
     PY: ['South America', 'AMER'],
     QA: ['Western Asia', 'EMEA'],
-    RE: ['Eastern Africa', 'AMER'],
+    RE: ['Eastern Africa', 'EMEA'],
     RO: ['Eastern Europe', 'EMEA'],
     RS: ['South-Eastern Europe', 'EMEA'],
     RU: ['Eastern Europe', 'EMEA'],
@@ -731,11 +915,10 @@ E.regions = {
     SB: ['Melanesia', 'APAC'],
     SC: ['Eastern Africa', 'APAC'],
     SD: ['Northern Africa', 'EMEA'],
-    SE: ['Scandinavia', 'Northern Europe', 'EMEA'],
+    SE: ['Scandinavia', 'EMEA', 'Northern Europe'],
     SG: ['South-Eastern Asia', 'APAC'],
     SH: ['Western Africa', 'AMER'],
     SI: ['South-Eastern Europe', 'EMEA'],
-    SJ: ['Scandinavia', 'Northern Europe', 'AMER'],
     SK: ['Eastern Europe', 'EMEA'],
     SL: ['Western Africa', 'EMEA'],
     SM: ['Southern Europe', 'EMEA'],
@@ -745,7 +928,7 @@ E.regions = {
     SS: ['Eastern Africa', 'EMEA'],
     ST: ['Middle Africa', 'EMEA'],
     SV: ['Central America', 'AMER'],
-    SX: ['Caribbean', 'APAC'],
+    SX: ['Caribbean', 'AMER'],
     SY: ['Western Asia', 'EMEA'],
     SZ: ['Southern Africa', 'EMEA'],
     TC: ['Caribbean', 'AMER'],
@@ -759,7 +942,6 @@ E.regions = {
     TM: ['Central Asia', 'EMEA'],
     TN: ['Northern Africa', 'EMEA'],
     TO: ['Polynesia', 'APAC'],
-    TP: ['South-Eastern Asia'],
     TR: ['South-Eastern Europe', 'EMEA'],
     TT: ['Caribbean', 'AMER'],
     TV: ['Polynesia', 'AMER'],
@@ -768,7 +950,6 @@ E.regions = {
     UA: ['Eastern Europe', 'EMEA'],
     UG: ['Eastern Africa', 'EMEA'],
     UK: ['Northern Europe'],
-    UM: ['Northern America', 'AMER'],
     US: ['Northern America', 'AMER'],
     UY: ['South America', 'AMER'],
     UZ: ['Central Asia', 'EMEA'],
@@ -781,10 +962,9 @@ E.regions = {
     VU: ['Melanesia', 'APAC'],
     WF: ['Polynesia', 'AMER'],
     WS: ['Polynesia', 'APAC'],
-    XK: ['South-Eastern Europe'],
+    XK: ['South-Eastern Europe', 'EMEA'],
     YE: ['Western Asia', 'EMEA'],
     YT: ['Eastern Africa', 'EMEA'],
-    YU: ['South-Eastern Europe'],
     ZA: ['Southern Africa', 'EMEA'],
     ZM: ['Eastern Africa', 'EMEA'],
     ZW: ['Eastern Africa', 'EMEA'],
@@ -898,12 +1078,9 @@ E.dialing_code_list = {
     BM: '1441',
     BT: '975',
     BO: '591',
-    BQ: '599',
     BA: '387',
     BW: '267',
-    BV: '55',
     BR: '55',
-    IO: '246',
     BN: '673',
     BG: '359',
     BF: '226',
@@ -933,7 +1110,6 @@ E.dialing_code_list = {
     DK: '45',
     DJ: '253',
     DM: '1767',
-    TP: '670',
     EC: '593',
     EG: '20',
     SV: '503',
@@ -946,8 +1122,6 @@ E.dialing_code_list = {
     FJ: '679',
     FI: '358',
     FR: '33',
-    FX: '33',
-    GF: '594',
     PF: '689',
     TF: '262',
     GA: '241',
@@ -960,7 +1134,6 @@ E.dialing_code_list = {
     GR: '30',
     GL: '299',
     GD: '1473',
-    GP: '590',
     GU: '1671',
     GT: '502',
     GG: '44',
@@ -985,7 +1158,6 @@ E.dialing_code_list = {
     JO: '962',
     KE: '254',
     KI: '686',
-    XK: '383',
     KW: '965',
     KG: '996',
     LA: '856',
@@ -1048,7 +1220,6 @@ E.dialing_code_list = {
     PL: '48',
     PT: '351',
     QA: '974',
-    RE: '262',
     RO: '40',
     RW: '250',
     BL: '590',
@@ -1056,10 +1227,8 @@ E.dialing_code_list = {
     KN: '1869',
     LC: '1758',
     MF: '590',
-    PM: '508',
     VC: '1784',
     SM: '378',
-    SX: '1721',
     ST: '239',
     SA: '966',
     SN: '221',
@@ -1079,7 +1248,6 @@ E.dialing_code_list = {
     LK: '94',
     SD: '249',
     SR: '597',
-    SJ: '47',
     SZ: '268',
     SE: '46',
     CH: '41',
@@ -1114,17 +1282,547 @@ E.dialing_code_list = {
     EH: '212',
     WS: '685',
     YE: '967',
-    YU: '38',
     ZM: '260',
     ZW: '263',
     DO: '1',
     CA: '1',
-    HM: '0',
     KZ: '7',
     US: '1',
-    UM: '1',
     PR: '1',
     RU: '7',
+};
+
+E.dialing_code_rlist = {
+    1: 'US',
+    1403: 'CA', 1587: 'CA', 1780: 'CA', 1825: 'CA', // Alberta
+    1236: 'CA', 1250: 'CA', 1604: 'CA', 1778: 'CA', // British Columbia
+    1204: 'CA', 1431: 'CA', // Manitoba
+    1506: 'CA', // New Brunswick
+    1709: 'CA', // Newfoundland and Labrador
+    1902: 'CA', 1782: 'CA', // Nova Scotia and Prince Edward Island
+    1226: 'CA', 1249: 'CA', 1289: 'CA', 1343: 'CA', 1365: 'CA', 1416: 'CA',
+    1437: 'CA', 1519: 'CA', 1548: 'CA', 1613: 'CA', 1647: 'CA', 1705: 'CA',
+    1807: 'CA', 1905: 'CA', // Ontario
+    1418: 'CA', 1438: 'CA', 1450: 'CA', 1514: 'CA', 1579: 'CA', 1581: 'CA',
+    1819: 'CA', 1873: 'CA', // Quebec
+    1306: 'CA', 1639: 'CA', // Saskatchewan
+    1867: 'CA', // Yukon, Northwest Territories and Nunavut
+    1787: 'PR', 1939: 'PR',
+    1809: 'DO', 1829: 'DO', 1849: 'DO',
+    1242: 'BS',
+    1246: 'BB',
+    1264: 'AI',
+    1268: 'AG',
+    1284: 'VG',
+    1340: 'VI',
+    1345: 'KY',
+    1441: 'BM',
+    1473: 'GD',
+    1649: 'TC',
+    1664: 'MS',
+    1670: 'MP',
+    1671: 'GU',
+    1684: 'AS',
+    1758: 'LC',
+    1767: 'DM',
+    1784: 'VC',
+    1868: 'TT',
+    1869: 'KN',
+    1876: 'JM',
+    20: 'EG',
+    211: 'SS',
+    212: 'MA',
+    212528: 'EH',
+    213: 'DZ',
+    216: 'TN',
+    218: 'LY',
+    220: 'GM',
+    221: 'SN',
+    222: 'MR',
+    223: 'ML',
+    224: 'GN',
+    225: 'CI',
+    226: 'BF',
+    227: 'NE',
+    228: 'TG',
+    229: 'BJ',
+    230: 'MU',
+    231: 'LR',
+    232: 'SL',
+    233: 'GH',
+    234: 'NG',
+    235: 'TD',
+    236: 'CF',
+    237: 'CM',
+    238: 'CV',
+    239: 'ST',
+    240: 'GQ',
+    241: 'GA',
+    242: 'CG',
+    243: 'CD',
+    244: 'AO',
+    245: 'GW',
+    248: 'SC',
+    249: 'SD',
+    250: 'RW',
+    251: 'ET',
+    252: 'SO',
+    253: 'DJ',
+    254: 'KE',
+    255: 'TZ',
+    256: 'UG',
+    257: 'BI',
+    258: 'MZ',
+    260: 'ZM',
+    261: 'MG',
+    262: 'TF',
+    262269: 'YT', 262639: 'YT',
+    263: 'ZW',
+    264: 'NA',
+    265: 'MW',
+    266: 'LS',
+    267: 'BW',
+    268: 'SZ',
+    269: 'KM',
+    27: 'ZA',
+    290: 'SH',
+    291: 'ER',
+    297: 'AW',
+    298: 'FO',
+    299: 'GL',
+    30: 'GR',
+    31: 'NL',
+    32: 'BE',
+    33: 'FR',
+    34: 'ES',
+    350: 'GI',
+    351: 'PT',
+    352: 'LU',
+    353: 'IE',
+    354: 'IS',
+    355: 'AL',
+    356: 'MT',
+    357: 'CY',
+    358: 'FI',
+    35818: 'AX', 358457: 'AX',
+    359: 'BG',
+    36: 'HU',
+    370: 'LT',
+    371: 'LV',
+    372: 'EE',
+    373: 'MD',
+    374: 'AM',
+    375: 'BY',
+    376: 'AD',
+    377: 'MC',
+    378: 'SM',
+    380: 'UA',
+    381: 'RS',
+    382: 'ME',
+    385: 'HR',
+    386: 'SI',
+    387: 'BA',
+    389: 'MK',
+    39: 'IT',
+    39066987: 'VA',
+    39066988: 'VA',
+    40: 'RO',
+    41: 'CH',
+    420: 'CZ',
+    421: 'SK',
+    423: 'LI',
+    43: 'AT',
+    44: 'GB',
+    441481: 'GG',
+    441534: 'JE',
+    441624: 'IM',
+    45: 'DK',
+    46: 'SE',
+    47: 'NO',
+    48: 'PL',
+    49: 'DE',
+    500: 'FK',
+    501: 'BZ',
+    502: 'GT',
+    503: 'SV',
+    504: 'HN',
+    505: 'NI',
+    506: 'CR',
+    507: 'PA',
+    509: 'HT',
+    51: 'PE',
+    52: 'MX',
+    53: 'CU',
+    54: 'AR',
+    55: 'BR',
+    56: 'CL',
+    57: 'CO',
+    58: 'VE',
+    590: 'BL',
+    591: 'BO',
+    592: 'GY',
+    593: 'EC',
+    595: 'PY',
+    596: 'MQ',
+    597: 'SR',
+    598: 'UY',
+    599: 'AN',
+    60: 'MY',
+    61: 'AU',
+    619162: 'CC',
+    619164: 'CX',
+    62: 'ID',
+    63: 'PH',
+    64: 'NZ',
+    65: 'SG',
+    66: 'TH',
+    670: 'TL',
+    6721: 'AQ',
+    67232: 'NF',
+    67235: 'NF',
+    673: 'BN',
+    674: 'NR',
+    675: 'PG',
+    676: 'TO',
+    677: 'SB',
+    678: 'VU',
+    679: 'FJ',
+    680: 'PW',
+    681: 'WF',
+    682: 'CK',
+    683: 'NU',
+    685: 'WS',
+    686: 'KI',
+    687: 'NC',
+    688: 'TV',
+    689: 'PF',
+    690: 'TK',
+    691: 'FM',
+    692: 'MH',
+    7: 'RU',
+    76: 'KZ', 77: 'KZ',
+    81: 'JP',
+    82: 'KR',
+    84: 'VN',
+    850: 'KP',
+    852: 'HK',
+    853: 'MO',
+    855: 'KH',
+    856: 'LA',
+    86: 'CN',
+    880: 'BD',
+    886: 'TW',
+    90: 'TR',
+    91: 'IN',
+    92: 'PK',
+    93: 'AF',
+    94: 'LK',
+    95: 'MM',
+    960: 'MV',
+    961: 'LB',
+    962: 'JO',
+    963: 'SY',
+    964: 'IQ',
+    965: 'KW',
+    966: 'SA',
+    967: 'YE',
+    968: 'OM',
+    970: 'PS',
+    971: 'AE',
+    972: 'IL',
+    973: 'BH',
+    974: 'QA',
+    975: 'BT',
+    976: 'MN',
+    977: 'NP',
+    98: 'IR',
+    992: 'TJ',
+    993: 'TM',
+    994: 'AZ',
+    995: 'GE',
+    996: 'KG',
+    998: 'UZ',
+};
+
+// records with "Parent" column equal to "Country"/"Region" from this list:
+// https://developers.google.com/adwords/api/docs/appendix/geotargeting
+// intersected with "Region Setting" from https://www.google.com/preferences
+// plus countries from https://support.google.com/google-ads/answer/6163740
+E.google_list = {
+    AD: 'Andorra',
+    AE: 'United Arab Emirates',
+    AF: 'Afghanistan',
+    AG: 'Antigua and Barbuda',
+    AI: 'Anguilla',
+    AL: 'Albania',
+    AM: 'Armenia',
+    AO: 'Angola',
+    AQ: 'Antarctica',
+    AR: 'Argentina',
+    AS: 'American Samoa',
+    AT: 'Austria',
+    AU: 'Australia',
+    AZ: 'Azerbaijan',
+    BA: 'Bosnia and Herzegovina',
+    BB: 'Barbados',
+    BD: 'Bangladesh',
+    BE: 'Belgium',
+    BF: 'Burkina Faso',
+    BG: 'Bulgaria',
+    BH: 'Bahrain',
+    BI: 'Burundi',
+    BJ: 'Benin',
+    BN: 'Brunei',
+    BO: 'Bolivia',
+    BQ: 'Caribbean Netherlands',
+    BR: 'Brazil',
+    BS: 'The Bahamas',
+    BT: 'Bhutan',
+    BW: 'Botswana',
+    BY: 'Belarus',
+    BZ: 'Belize',
+    CA: 'Canada',
+    CC: 'Cocos (Keeling) Islands',
+    CD: 'Democratic Republic of the Congo',
+    CF: 'Central African Republic',
+    CG: 'Republic of the Congo',
+    CH: 'Switzerland',
+    CI: 'Cote d\'Ivoire',
+    CK: 'Cook Islands',
+    CL: 'Chile',
+    CM: 'Cameroon',
+    CN: 'China',
+    CO: 'Colombia',
+    CR: 'Costa Rica',
+    CU: 'Cuba',
+    CV: 'Cape Verde',
+    CW: 'Curacao',
+    CX: 'Christmas Island',
+    CY: 'Cyprus',
+    CZ: 'Czechia',
+    DE: 'Germany',
+    DJ: 'Djibouti',
+    DK: 'Denmark',
+    DM: 'Dominica',
+    DO: 'Dominican Republic',
+    DZ: 'Algeria',
+    EC: 'Ecuador',
+    EE: 'Estonia',
+    EG: 'Egypt',
+    ER: 'Eritrea',
+    ES: 'Spain',
+    ET: 'Ethiopia',
+    FI: 'Finland',
+    FJ: 'Fiji',
+    FM: 'Federated States of Micronesia',
+    FR: 'France',
+    GA: 'Gabon',
+    GB: 'United Kingdom',
+    GD: 'Grenada',
+    GE: 'Georgia',
+    GG: 'Guernsey',
+    GH: 'Ghana',
+    GI: 'Gibraltar',
+    GL: 'Greenland',
+    GM: 'The Gambia',
+    GN: 'Guinea',
+    GQ: 'Equatorial Guinea',
+    GR: 'Greece',
+    GS: 'South Georgia and the South Sandwich Islands',
+    GT: 'Guatemala',
+    GU: 'Guam',
+    GW: 'Guinea-Bissau',
+    GY: 'Guyana',
+    HK: 'Hong Kong',
+    HM: 'Heard Island and McDonald Islands',
+    HN: 'Honduras',
+    HR: 'Croatia',
+    HT: 'Haiti',
+    HU: 'Hungary',
+    ID: 'Indonesia',
+    IE: 'Ireland',
+    IL: 'Israel',
+    IM: 'Isle of Man,United Kingdom',
+    IN: 'India',
+    IQ: 'Iraq',
+    IR: 'Iran',
+    IS: 'Iceland',
+    IT: 'Italy',
+    JE: 'Jersey',
+    JM: 'Jamaica',
+    JO: 'Jordan',
+    JP: 'Japan',
+    KE: 'Kenya',
+    KG: 'Kyrgyzstan',
+    KH: 'Cambodia',
+    KI: 'Kiribati',
+    KM: 'Comoros',
+    KN: 'Saint Kitts and Nevis',
+    KP: 'North Korea',
+    KR: 'South Korea',
+    KW: 'Kuwait',
+    KZ: 'Kazakhstan',
+    LA: 'Laos',
+    LB: 'Lebanon',
+    LC: 'Saint Lucia',
+    LI: 'Liechtenstein',
+    LK: 'Sri Lanka',
+    LR: 'Liberia',
+    LS: 'Lesotho',
+    LT: 'Lithuania',
+    LU: 'Luxembourg',
+    LV: 'Latvia',
+    LY: 'Libya',
+    MA: 'Morocco',
+    MC: 'Monaco',
+    MD: 'Moldova',
+    ME: 'Montenegro',
+    MG: 'Madagascar',
+    MH: 'Marshall Islands',
+    MK: 'North Macedonia',
+    ML: 'Mali',
+    MM: 'Myanmar (Burma)',
+    MN: 'Mongolia',
+    MP: 'Northern Mariana Islands',
+    MR: 'Mauritania',
+    MS: 'Montserrat',
+    MT: 'Malta',
+    MU: 'Mauritius',
+    MV: 'Maldives',
+    MW: 'Malawi',
+    MX: 'Mexico',
+    MY: 'Malaysia',
+    MZ: 'Mozambique',
+    NA: 'Namibia',
+    NC: 'New Caledonia',
+    NE: 'Niger',
+    NF: 'Norfolk Island',
+    NG: 'Nigeria',
+    NI: 'Nicaragua',
+    NL: 'Netherlands',
+    NO: 'Norway',
+    NP: 'Nepal',
+    NR: 'Nauru',
+    NU: 'Niue',
+    NZ: 'New Zealand',
+    OM: 'Oman',
+    PA: 'Panama',
+    PE: 'Peru',
+    PF: 'French Polynesia',
+    PG: 'Papua New Guinea',
+    PH: 'Philippines',
+    PK: 'Pakistan',
+    PL: 'Poland',
+    PM: 'Saint Pierre and Miquelon',
+    PN: 'Pitcairn Islands',
+    PR: 'Puerto Rico',
+    PS: 'Palestine',
+    PT: 'Portugal',
+    PW: 'Palau',
+    PY: 'Paraguay',
+    QA: 'Qatar',
+    RO: 'Romania',
+    RS: 'Serbia',
+    RU: 'Russia',
+    RW: 'Rwanda',
+    SA: 'Saudi Arabia',
+    SB: 'Solomon Islands',
+    SC: 'Seychelles',
+    SD: 'Sudan',
+    SE: 'Sweden',
+    SG: 'Singapore',
+    SH: 'Saint Helena, Ascension and Tristan da Cunha',
+    SI: 'Slovenia',
+    SK: 'Slovakia',
+    SL: 'Sierra Leone',
+    SM: 'San Marino',
+    SN: 'Senegal',
+    SO: 'Somalia',
+    SR: 'Suriname',
+    ST: 'Sao Tome and Principe',
+    SV: 'El Salvador',
+    SX: 'Sint Maarten',
+    SY: 'Syria',
+    SZ: 'Eswatini',
+    TD: 'Chad',
+    TF: 'French Southern and Antarctic Lands',
+    TG: 'Togo',
+    TH: 'Thailand',
+    TJ: 'Tajikistan',
+    TK: 'Tokelau',
+    TL: 'Timor-Leste',
+    TM: 'Turkmenistan',
+    TN: 'Tunisia',
+    TO: 'Tonga',
+    TR: 'Turkey',
+    TT: 'Trinidad and Tobago',
+    TV: 'Tuvalu',
+    TW: 'Taiwan',
+    TZ: 'Tanzania',
+    UA: 'Ukraine',
+    UG: 'Uganda',
+    UK: 'United Kingdom',
+    UM: 'United States Minor Outlying Islands',
+    US: 'United States',
+    UY: 'Uruguay',
+    UZ: 'Uzbekistan',
+    VA: 'Vatican City',
+    VC: 'Saint Vincent and the Grenadines',
+    VE: 'Venezuela',
+    VG: 'British Virgin Islands',
+    VI: 'U.S. Virgin Islands',
+    VN: 'Vietnam',
+    VU: 'Vanuatu',
+    WF: 'Wallis and Futuna',
+    WS: 'Samoa',
+    YE: 'Yemen',
+    ZA: 'South Africa',
+    ZM: 'Zambia',
+    ZW: 'Zimbabwe',
+};
+
+E.currency = {
+    AD: 'EUR', AE: 'AED', AF: 'AFN', AG: 'XCD', AI: 'XCD', AL: 'ALL',
+    AM: 'AMD', AN: 'ANG', AO: 'AOA', AR: 'ARS', AS: 'USD', AT: 'EUR',
+    AU: 'AUD', AW: 'AWG', AX: 'EUR', AZ: 'AZN', BA: 'BAM', BB: 'BBD',
+    BD: 'BDT', BE: 'EUR', BF: 'XOF', BG: 'BGN', BH: 'BHD', BI: 'BIF',
+    BJ: 'XOF', BL: 'EUR', BM: 'BMD', BN: 'BND', BO: 'BOB', BQ: 'USD',
+    BR: 'BRL', BS: 'BSD', BT: 'BTN', BW: 'BWP', BY: 'BYN', BZ: 'BZD',
+    CA: 'CAD', CC: 'AUD', CD: 'CDF', CF: 'XAF', CG: 'XAF', CH: 'CHF',
+    CI: 'XOF', CK: 'NZD', CL: 'CLP', CM: 'XAF', CN: 'CNY', CO: 'COP',
+    CR: 'CRC', CU: 'CUP', CV: 'CVE', CW: 'ANG', CX: 'AUD', CY: 'EUR',
+    CZ: 'CZK', DE: 'EUR', DJ: 'DJF', DK: 'DKK', DM: 'XCD', DO: 'DOP',
+    DZ: 'DZD', EC: 'USD', EE: 'EUR', EG: 'EGP', EH: 'MAD', ER: 'ERN',
+    ES: 'EUR', ET: 'ETB', EU: 'EUR', FI: 'EUR', FJ: 'FJD', FK: 'FKP',
+    FM: 'USD', FO: 'DKK', FR: 'EUR', GA: 'XAF', GB: 'GBP', GD: 'XCD',
+    GE: 'GEL', GF: 'EUR', GG: 'GBP', GH: 'GHS', GI: 'GIP', GL: 'DKK',
+    GM: 'GMD', GN: 'GNF', GP: 'EUR', GQ: 'XAF', GR: 'EUR', GS: 'FKP',
+    GT: 'GTQ', GU: 'USD', GW: 'XOF', GY: 'GYD', HK: 'HKD', HN: 'HNL',
+    HR: 'HRK', HT: 'HTG', HU: 'HUF', ID: 'IDR', IE: 'EUR', IL: 'ILS',
+    IM: 'GBP', IN: 'INR', IQ: 'IQD', IR: 'IRR', IS: 'ISK', IT: 'EUR',
+    JE: 'GBP', JM: 'JMD', JO: 'JOD', JP: 'JPY', KE: 'KES', KG: 'KGS',
+    KH: 'KHR', KI: 'AUD', KM: 'KMF', KN: 'XCD', KP: 'KPW', KR: 'KRW',
+    KW: 'KWD', KY: 'KYD', KZ: 'KZT', LA: 'LAK', LB: 'LBP', LC: 'XCD',
+    LI: 'CHF', LK: 'LKR', LR: 'LRD', LS: 'LSL', LT: 'EUR', LU: 'EUR',
+    LV: 'EUR', LY: 'LYD', MA: 'MAD', MC: 'EUR', MD: 'MDL', ME: 'EUR',
+    MF: 'EUR', MG: 'MGA', MH: 'USD', MK: 'MKD', ML: 'XOF', MM: 'MMK',
+    MN: 'MNT', MO: 'MOP', MP: 'USD', MQ: 'EUR', MR: 'MRU', MS: 'XCD',
+    MT: 'EUR', MU: 'MUR', MV: 'MVR', MW: 'MWK', MX: 'MXN', MY: 'MYR',
+    MZ: 'MZN', NA: 'NAD', NC: 'XPF', NE: 'XOF', NF: 'AUD', NG: 'NGN',
+    NI: 'NIO', NL: 'EUR', NO: 'NOK', NP: 'NPR', NR: 'AUD', NU: 'NZD',
+    NZ: 'NZD', OM: 'OMR', PA: 'PAB', PE: 'PEN', PF: 'XPF', PG: 'PGK',
+    PH: 'PHP', PK: 'PKR', PL: 'PLN', PN: 'NZD', PR: 'USD', PS: 'ILS',
+    PT: 'EUR', PW: 'USD', PY: 'PYG', QA: 'QAR', RE: 'EUR', RO: 'RON',
+    RS: 'RSD', RU: 'RUB', RW: 'RWF', SA: 'SAR', SB: 'SBD', SC: 'SCR',
+    SD: 'SDG', SE: 'SEK', SG: 'SGD', SH: 'SHP', SI: 'EUR', SK: 'EUR',
+    SL: 'SLL', SM: 'EUR', SN: 'XOF', SO: 'SOS', SR: 'SRD', SS: 'SSP',
+    ST: 'STN', SV: 'USD', SX: 'ANG', SY: 'SYP', SZ: 'SZL', TC: 'USD',
+    TD: 'XAF', TF: 'EUR', TG: 'XOF', TH: 'THB', TJ: 'TJS', TK: 'NZD',
+    TL: 'USD', TM: 'TMT', TN: 'TND', TO: 'TOP', TR: 'TRY', TT: 'TTD',
+    TV: 'AUD', TW: 'TWD', TZ: 'TZS', UA: 'UAH', UG: 'UGX', UK: 'GBP',
+    US: 'USD', UY: 'UYU', UZ: 'UZS', VA: 'EUR', VC: 'XCD', VE: 'VES',
+    VG: 'GBP', VI: 'USD', VN: 'VND', VU: 'VUV', WF: 'XPF', WS: 'WST',
+    XK: 'EUR', YE: 'YER', YT: 'EUR', ZA: 'ZAR', ZM: 'ZMW', ZW: 'USD',
 };
 
 E.code2label = function(code){
@@ -1173,5 +1871,10 @@ E.www2code = function(code){
 
 E.code2www = function(code){
     return code=='GB' ? 'UK' : code=='gb' ? 'uk' : code; };
+
+E.gl2uule = function(gl){
+    gl = E.www2code(gl.toUpperCase());
+    return E.google_list[gl] || gl;
+};
 
 return E; }); }());
