@@ -28,7 +28,7 @@ E.thousand_grouping = function(num_s){
     var m = /^([-+])?(\d*)(\.\d*)?$/.exec(num_s);
     if (!m)
         return num_s;
-    m[2] = (m[2]||'').replace(/(\d)(?=(\d{3})+(?!\d))/g, '$1'+',');
+    m[2] = (m[2]||'').replace(/(\d)(?=(\d{3})+(?!\d))/g, '$1,');
     return (m[1]||'')+m[2]+(m[3]||'');
 };
 
@@ -237,121 +237,146 @@ E.parse_slow = function(fmt){
     var _f = [], out, arg, arg_s, argv, sign;
     function f(fn){ _f.push(fn); }
     for (; _fmt; _fmt = _fmt.substring(match[0].length))
-    (function(){
-        if (match = /^[^%]+/.exec(_fmt))
-        {
-            var _match = match;
-            f(function(){ return out += _match[0]; });
-        }
-        else if (match = /^%%/.exec(_fmt))
-            f(function(){ return out += '%'; });
-        else if ((match =
-            /^%(?:([1-9]\d*)\$|\(([^\)]+)\))?(\+)?(0|'[^$])?(-)?(\d+)?(?:\.(\d+))?(')?([bcdefoOsuxX])/
-            .exec(_fmt)))
-        {
-            var positional = match[1], keyword = match[2], sign = match[3];
-            var pad_zero = match[4], pad_min = match[5], pad_max = match[6];
-            var precision = match[7], thousand_grouping = match[8]=="'";
-            var conversion = match[9], keyword_list = [], _cursor = cursor;
-            if (keyword)
+    {
+        (function(){
+            if (match = /^[^%]+/.exec(_fmt))
             {
-                arg_names |= 1;
-                var _keyword = keyword, kmatch;
-                if (!(kmatch = /^([a-z_][a-z_\d]*)/i.exec(_keyword)))
-                    throw 'sprintf: invalid keyword property name '+_keyword;
-                keyword_list.push(kmatch[1]);
-                while (_keyword = _keyword.substring(kmatch[0].length))
+                var _match = match;
+                f(function(){ return out += _match[0]; });
+            }
+            else if (match = /^%%/.exec(_fmt))
+                f(function(){ return out += '%'; });
+            else if ((match =
+                /^%(?:([1-9]\d*)\$|\(([^\)]+)\))?(\+)?(0|'[^$])?(-)?(\d+)?(?:\.(\d+))?(')?([bcdefoOsuxX])/
+                .exec(_fmt)))
+            {
+                var positional = match[1], keyword = match[2], sign = match[3];
+                var pad_zero = match[4], pad_min = match[5];
+                var pad_max = match[6];
+                var precision = match[7], thousand_grouping = match[8]=="'";
+                var conversion = match[9], keyword_list = [], _cursor = cursor;
+                if (keyword)
                 {
-                    if (kmatch = /^\.([a-z_][a-z_\d]*)/i.exec(_keyword))
-                        keyword_list.push(kmatch[1]);
-                    else if (kmatch = /^\[(\d+)\]/.exec(_keyword))
-                        keyword_list.push(kmatch[1]);
-                    else
-                        throw 'sprintf: invalid keyword format '+_keyword;
+                    arg_names |= 1;
+                    var _keyword = keyword, kmatch;
+                    if (!(kmatch = /^([a-z_][a-z_\d]*)/i.exec(_keyword)))
+                    {
+                        throw 'sprintf: invalid keyword property name '
+                            +_keyword;
+                    }
+                    keyword_list.push(kmatch[1]);
+                    while (_keyword = _keyword.substring(kmatch[0].length))
+                    {
+                        if (kmatch = /^\.([a-z_][a-z_\d]*)/i.exec(_keyword))
+                            keyword_list.push(kmatch[1]);
+                        else if (kmatch = /^\[(\d+)\]/.exec(_keyword))
+                            keyword_list.push(kmatch[1]);
+                        else
+                            throw 'sprintf: invalid keyword format '+_keyword;
+                    }
                 }
-            }
-            else
-                arg_names |= 2;
-            if (arg_names===3)
-            {
-                throw 'sprintf: mixing positional and named placeholders is '
-                    +'not (yet) supported';
-            }
-            f(function(){ sign = false; });
-            if (keyword_list.length) // keyword argument
-            {
+                else
+                    arg_names |= 2;
+                if (arg_names===3)
+                {
+                    throw 'sprintf: mixing positional and named placeholders'
+                        +' is not (yet) supported';
+                }
+                f(function(){ sign = false; });
+                if (keyword_list.length) // keyword argument
+                {
+                    f(function(){
+                        arg = argv[_cursor];
+                        for (var k = 0; k < keyword_list.length && arg!=null;
+                            k++)
+                        {
+                            arg = arg[keyword_list[k]];
+                        }
+                    });
+                }
+                else if (positional) // positional argument (explicit)
+                    f(function(){ arg = argv[positional]; });
+                else // positional argument (implicit)
+                {
+                    f(function(){ arg = argv[_cursor]; });
+                    cursor++;
+                }
+                if (/[^sO]/.test(conversion))
+                    f(function(){ return arg = +arg; });
+                switch (conversion)
+                {
+                case 'b': f(function(){ arg_s = arg.toString(2); }); break;
+                case 'c':
+                      f(function(){ arg_s = String.fromCharCode(arg); });
+                      break;
+                case 'd':
+                    f(function(){
+                        arg = sprintf.to_int(arg); arg_s = ''+arg; });
+                    if (thousand_grouping)
+                    {
+                        f(function(){
+                            arg_s = sprintf.thousand_grouping(arg_s); });
+                    }
+                    break;
+                case 'e':
+                    f(function(){ arg_s = arg.toExponential(
+                        precision ? precision : undefined); });
+                    break;
+                case 'f':
+                    if (precision)
+                        f(function(){ arg_s = arg.toFixed(precision); });
+                    else
+                        f(function(){ arg_s = ''+arg; });
+                    if (thousand_grouping)
+                    {
+                        f(function(){
+                            arg_s = sprintf.thousand_grouping(arg_s); });
+                    }
+                    break;
+                case 'o': f(function(){ arg_s = arg.toString(8); }); break;
+                case 'O': f(function(){ arg_s = stringify(arg); }); break;
+                case 'u': f(function(){ arg = arg >>> 0; arg_s = ''+arg; });
+                          break;
+                case 'x': f(function(){ arg_s = arg.toString(16); }); break;
+                case 'X':
+                    f(function(){ arg_s = arg.toString(16).toUpperCase(); });
+                    break;
+                case 's':
+                    f(function(){ arg_s = ''+arg; });
+                    if (precision)
+                    {
+                        f(function(){
+                            arg_s = arg_s.substring(0, precision); });
+                    }
+                    break;
+                }
+                if (/[def]/.test(conversion))
+                {
+                    if (sign)
+                    {
+                        f(function(){
+                            if (arg>=0)
+                                arg_s = '+'+arg_s; });
+                    }
+                    f(function(){ sign = arg_s[0]=='-' || arg_s[0]=='+'; });
+                }
+                var pad_chr = !pad_zero ? ' ' : pad_zero=='0' ? '0' :
+                    pad_zero[1];
                 f(function(){
-                    arg = argv[_cursor];
-                    for (var k = 0; k < keyword_list.length && arg!=null; k++)
-                        arg = arg[keyword_list[k]];
+                    var pad_chrs = pad_chr.repeat(
+                        Math.max(+pad_max-arg_s.length, 0));
+                    var arg_padded = !pad_max ? arg_s :
+                        pad_min ? arg_s+pad_chrs :
+                        sign && pad_chr[0]=='0' ?
+                        arg_s[0]+pad_chrs+arg_s.slice(1) :
+                        pad_chrs+arg_s;
+                    out += arg_padded;
                 });
             }
-            else if (positional) // positional argument (explicit)
-                f(function(){ arg = argv[positional]; });
-            else // positional argument (implicit)
-            {
-                f(function(){ arg = argv[_cursor]; });
-                cursor++;
-            }
-            if (/[^sO]/.test(conversion))
-                f(function(){ return arg = +arg; });
-            switch (conversion)
-            {
-            case 'b': f(function(){ arg_s = arg.toString(2); }); break;
-            case 'c':
-                  f(function(){ arg_s = String.fromCharCode(arg); });
-                  break;
-            case 'd':
-                f(function(){ arg = sprintf.to_int(arg); arg_s = ''+arg; });
-                if (thousand_grouping)
-                    f(function(){ arg_s = sprintf.thousand_grouping(arg_s); });
-                break;
-            case 'e':
-                f(function(){ arg_s = arg.toExponential(
-                    precision ? precision : undefined); });
-                break;
-            case 'f':
-                if (precision)
-                    f(function(){ arg_s = arg.toFixed(precision); });
-                else
-                    f(function(){ arg_s = ''+arg; });
-                if (thousand_grouping)
-                    f(function(){ arg_s = sprintf.thousand_grouping(arg_s); });
-                break;
-            case 'o': f(function(){ arg_s = arg.toString(8); }); break;
-            case 'O': f(function(){ arg_s = stringify(arg); }); break;
-            case 'u': f(function(){ arg = arg >>> 0; arg_s = ''+arg; }); break;
-            case 'x': f(function(){ arg_s = arg.toString(16); }); break;
-            case 'X':
-                f(function(){ arg_s = arg.toString(16).toUpperCase(); });
-                break;
-            case 's':
-                f(function(){ arg_s = ''+arg; });
-                if (precision)
-                    f(function(){ arg_s = arg_s.substring(0, precision); });
-                break;
-            }
-            if (/[def]/.test(conversion))
-            {
-                if (sign)
-                    f(function(){ if (arg>=0) arg_s = '+'+arg_s; });
-                f(function(){ sign = arg_s[0]=='-' || arg_s[0]=='+'; });
-            }
-            var pad_chr = !pad_zero ? ' ' : pad_zero=='0' ? '0' : pad_zero[1];
-            f(function(){
-                var pad_chrs = pad_chr.repeat(
-                    Math.max(+pad_max-arg_s.length, 0));
-                var arg_padded = !pad_max ? arg_s :
-                    pad_min ? arg_s+pad_chrs :
-                    sign && pad_chr[0]=='0' ?
-                    arg_s[0]+pad_chrs+arg_s.slice(1) :
-                    pad_chrs+arg_s;
-                out += arg_padded;
-            });
-        }
-        else
-            throw 'sprintf invalid format '+_fmt;
-    })();
+            else
+                throw 'sprintf invalid format '+_fmt;
+        })();
+    }
     return function(_argv){
         argv = _argv;
         out = '';
@@ -362,7 +387,7 @@ E.parse_slow = function(fmt){
 };
 E.parse = (function(){
     try {
-        if ((new Function('return 1')())==1)
+        if (new Function('return 1')()==1)
             return E.parse_fast;
     } catch(e){}
     return E.parse_slow; // capp does not support new Function()
