@@ -1,5 +1,5 @@
 // LICENSE_CODE ZON ISC
-'use strict'; /*zlint br*//*jslint node:true*/
+'use strict'; /*zlint br*//*jslint node:true, browser:true*/
 (function(){
 var define;
 var is_node = typeof module=='object' && module.exports;
@@ -8,13 +8,13 @@ if (!is_node)
 else
     define = function(){};
 
-define(['cookie'], function(cookie){
+define(['cookie', '/util/util.js'], function(cookie, zutil){
 var E = {};
 var storage;
 
 function have_local_storage(){
     try {
-        var _ = localStorage;
+        var _ = window.localStorage;
         if (_.length)
             return true;
         _.setItem('_', 0);
@@ -23,7 +23,7 @@ function have_local_storage(){
     } catch(e){}
 }
 
-function select_local_storage(){ storage = localStorage; }
+function select_local_storage(){ storage = window.localStorage; }
 
 function select_cookies(domain){
     var cookie_opt = {domain: '.'+domain, path: '/', expires: 30};
@@ -33,36 +33,52 @@ function select_cookies(domain){
     };
 }
 
-E.init = function(domain){
+E.init = function(opt){
+    var domain;
+    try { domain = document.location.hostname; }
+    catch(e){ domain = 'luminati.io'; }
+    if (typeof opt=='string')
+        domain = opt;
+    // XXX arik HACK: remove test_storage once all tests are fixed and we can
+    // enable it
+    if (E.is_test_storage = zutil.get(opt, 'test_storage') && zutil.is_mocha())
+        return E.test_storage = {};
     if (have_local_storage())
         return select_local_storage();
     console.error('cannot use localStorage, using cookies instead');
-    select_cookies(domain||'hola.org');
+    select_cookies(domain);
 };
 E.init();
 
 E.on_err = function(){};
 
+// XXX arik: add simple storage test
 E.set = function(key, val){
-    try { storage.setItem(key, val); }
-    catch(err){ E.on_err('storage_set', key, err); }
+    if (E.is_test_storage)
+        return E.test_storage[key] = val;
+    try { return storage.setItem(key, val); }
+    catch(e){ E.on_err('storage_set', key, e); }
 };
 
 E.get = function(key){
+    if (E.is_test_storage)
+        return E.test_storage[key];
     try { return storage.getItem(key); }
-    catch(err){ E.on_err('storage_get', key, err); }
+    catch(e){ E.on_err('storage_get', key, e); }
 };
 
 E.get_int = function(key){ return +E.get(key)||0; };
 
 E.clr = function(key){
+    if (E.is_test_storage)
+        return delete E.test_storage[key];
     try { storage.removeItem(key); }
-    catch(err){ E.on_err('storage_clr', key, err); }
+    catch(e){ E.on_err('storage_clr', key, e); }
 };
 
 E.set_json = function(key, val){
     try { return E.set(key, JSON.stringify(val||null)); }
-    catch(err){ E.on_err('storage_set_json', key, err); }
+    catch(e){ E.on_err('storage_set_json', key, e); }
 };
 
 E.get_json = function(key){
@@ -70,7 +86,7 @@ E.get_json = function(key){
     if (!val)
         return val;
     try { val = JSON.parse(val); }
-    catch(err){ console.log('err '+err); }
+    catch(e){ console.log('err '+e); }
     return val;
 };
 
