@@ -5,8 +5,6 @@ var define, process;
 var is_node = typeof module=='object' && module.exports && module.children;
 var is_rn = (typeof global=='object' && !!global.nativeRequire) ||
     (typeof navigator=='object' && navigator.product=='ReactNative');
-var is_ff_addon = typeof module=='object' && module.uri
-    && !module.uri.indexOf('resource://');
 if (is_rn)
 {
     define = require('./require_node.js').define(module, '../',
@@ -18,7 +16,7 @@ if (is_rn)
         env: {},
     };
 }
-else if (!is_node && !is_ff_addon)
+else if (!is_node)
 {
     define = self.define;
     process = {env: {}};
@@ -26,9 +24,7 @@ else if (!is_node && !is_ff_addon)
 else
 {
     define = require('./require_node.js').define(module, '../');
-    if (is_ff_addon)
-        process = {env: {}};
-    else if (is_node)
+    if (is_node)
     {
         process = global.process||require('_process');
         require('./config.js');
@@ -82,8 +78,6 @@ E.is = function(level){ return level<=E.level; };
     E.is[l] = function(){ return level<=E.level; };
 });
 
-E.log_tail = function(size){ return E.log.join('\n').substr(-(size||4096)); };
-
 /* perr is a stub overridden by upper layers */
 E.perr = function(id, info, opt){
     E._zerr(!opt || opt.level===undefined ? L.ERR : opt.level,
@@ -91,6 +85,8 @@ E.perr = function(id, info, opt){
     if (perr_pending && perr_pending.length<100)
         perr_pending.push(Array.from(arguments));
 };
+var perr_hooks = [];
+E.add_perr_hook = perr_hooks.push.bind(perr_hooks);
 var perr_dropped = {};
 var perr_orig = E.perr;
 function wrap_perr(perr_fn){
@@ -109,6 +105,8 @@ function wrap_perr(perr_fn){
         var rl = rl_hash[id] = rl_hash[id]||{};
         if (pre_send)
             pre_send(id, info, opt);
+        perr_hooks.filter(function(h){ return h.ids.test(id); })
+        .forEach(function(h){ h.fn(id, info, opt); });
         if (opt.rate_limit===false || rate_limit(rl, ms, count))
         {
             if (perr_dropped[id])
@@ -181,9 +179,42 @@ E.set_level = function(level){
     return prev;
 };
 
+E.get_stack_trace = function(opt){
+    if (!opt)
+        opt = {};
+    if (opt.limit===undefined)
+        opt.limit = Infinity;
+    if (opt.short===undefined)
+        opt.short = true;
+    var old_stack_limit = Error.stackTraceLimit;
+    if (opt.limit)
+        Error.stackTraceLimit = opt.limit;
+    var stack = zerr.e2s(new Error());
+    if (opt.limit)
+        Error.stackTraceLimit = old_stack_limit;
+    if (opt.short)
+    {
+        stack = stack
+            .replace(/^.+util\/etask.+$/gm, '    ...')
+            .replace(/( {4}\.\.\.\n)+/g, '    ...\n');
+    }
+    return stack;
+};
+
+E.log = [];
+E.log.max_size = 200;
+E.log_tail = function(size){
+    return (E.log||[]).join('\n').substr(-(size||4096)); };
+
+function log_tail_push(msg){
+    E.log.push(msg);
+    if (E.log.length>E.log.max_size)
+        E.log.splice(0, E.log.length - E.log.max_size/2);
+}
+
 if (is_node)
 { // zerr-node
-E.ZEXIT_LOG_DIR = '/tmp/zexit_logs';
+E.ZEXIT_LOG_DIR = env.ZEXIT_LOG_DIR||'/tmp/zexit_logs';
 E.prefix = '';
 
 E.level = L.NOTICE;
@@ -223,13 +254,18 @@ var __zerr = function(level, args){
     var msg = zerr_format(args);
     var k = Object.keys(L);
     var prefix = E.hide_timestamp ? '' : E.prefix+date.to_sql_ms()+' ';
-    console.error(prefix+k[level]+': '+msg);
+    if (env.CURRENT_SYSTEMD_UNIT_NAME)
+        prefix = '<'+level+'>'+prefix;
+    var res = prefix+k[level]+': '+msg;
+    console.error(res);
+    log_tail_push(res);
 };
 
 E.set_logger = function(logger){
     __zerr = function(level, args){
         var msg = zerr_format(args);
         logger(level, msg);
+        log_tail_push(E.prefix+date.to_sql_ms()+': '+msg);
     };
 };
 
@@ -252,8 +288,9 @@ E.zexit = function(args){
         var e = new Error();
         stack = e.stack;
         __zerr(L.CRIT, arguments);
-        console.error(stack);
     }
+    if ((args&&args.code)!='ERR_ASSERTION')
+        console.error('zerr.zexit was called', new Error().stack);
     E.flush();
     // workaround for process.zon override issue
     if (process.zon && process.zon.main)
@@ -332,14 +369,12 @@ _zerr = function(l, args){
         .substr(0, 1024);
         var prefix = (E.hide_timestamp ? '' : date.to_sql_ms()+' ')
         +L_STR[l]+': ';
-        E.log.push(prefix+s);
         if (E.is(l))
         {
             Function.prototype.apply.bind(console[console_method(l)],
                 console)([prefix+fmt].concat(fmt_args));
         }
-        if (E.log.length>E.log.max_size)
-            E.log.splice(0, E.log.length - E.log.max_size/2);
+        log_tail_push(prefix+s);
     } catch(err){
         try { console.error('ERROR in zerr '+(err.stack||err), arguments); }
         catch(e){}
